% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{as_any_event}
\alias{as_any_event}
\title{PathData to time-to-event data format for any state other than null state}
\usage{
as_any_event(pd, null_state = "Randomization")
}
\arguments{
\item{pd}{A \code{\link{PathData}} object}

\item{null_state}{Name of the base state}
}
\value{
A \code{\link{PathData}} object
}
\description{
PathData to time-to-event data format for any state other than null state
}
