% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{as_single_event}
\alias{as_single_event}
\title{PathData to single event format}
\usage{
as_single_event(pd, event)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object}

\item{event}{Name of the event of interest (character)}
}
\value{
A \code{\link{PathData}} object
}
\description{
PathData to single event format
}
