% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{bmstate-package}
\alias{bmstate-package}
\alias{bmstate}
\title{The 'bmstate' package.}
\description{
Bayesian multistate modeling using 'Stan'.
}
\section{Creating a model}{

A model is represented by a \code{\link{MultistateModel}}
object, which has a \code{\link{MultistateSystem}}, which has a
\code{\link{TransitionMatrix}}. You can create a model using
\code{\link{create_msm}}. For this,
you need to pass in a \code{\link{TransitionMatrix}} defining the
states and transitions. This can be created directly with
\code{TransitionMatrix$new()} or using built-in functions for common
multistate models such as \code{\link{transmat_illnessdeath}}.
}

\section{Fitting a model}{

You can fit a model with 'Stan' using \code{\link{fit_stan}}. You need
to pass in the model object and a \code{\link{JointData}} object, which
consists of \code{\link{PathData}} and \code{\link{DosingData}}. Unless
you are using the experimental feature of fitting a PK submodel, the
latter should be \code{NULL}.
}

\section{Prediction with a fitted model}{


\itemize{

\item For predicting state visit risks using state path simulation,
see the functions \code{\link{generate_paths}}, \code{\link{p_state_visit}},
and \code{\link{p_state_visit_per_subject}}.

\item For analytically solving state occupancy probabilities after some time,
  see \code{\link{solve_trans_prob_fit}}.
}
}

\section{Vignettes}{

See the vignettes for more information.
}

\author{
Juho Timonen (first.last at iki.fi) and Jacqueline Buros
}
\keyword{Bayesian}
\keyword{Stan}
\keyword{multistate}
