% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKModel.R
\name{compute_theta_pk}
\alias{compute_theta_pk}
\title{Compute PK params}
\usage{
compute_theta_pk(mu, beta, x)
}
\arguments{
\item{mu}{mean parameter values}

\item{beta}{values of covariate multipliers}

\item{x}{values of (normalized) covariates}
}
\description{
Compute PK params
}
