% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_brier_score_plot}
\alias{create_brier_score_plot}
\title{Plot Brier scores}
\usage{
create_brier_score_plot(ppsurv_subj, pd, sub_ids_char_training)
}
\arguments{
\item{ppsurv_subj}{\code{ppsurv} object}

\item{pd}{\code{\link{PathData}} object}

\item{sub_ids_char_training}{training subject ids (character)}
}
\description{
Plot Brier scores
}
