% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_cindex_plot}
\alias{create_cindex_plot}
\title{Plot concordance indices}
\usage{
create_cindex_plot(
  pd,
  paths_gen,
  paths_gen_oos,
  sub_ids_train,
  sub_ids_test,
  ppsurv_subj,
  ppsurv_subj_oos
)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object of observed data}

\item{paths_gen}{A \code{\link{PathData}} object of generated paths
(in-sample)}

\item{paths_gen_oos}{A \code{\link{PathData}} object of generated paths
(out-of-sample)}

\item{sub_ids_train}{training subject ids}

\item{sub_ids_test}{test subject ids}

\item{ppsurv_subj}{\code{ppsurv} object}

\item{ppsurv_subj_oos}{\code{ppsurv} object (out-of-sample)}
}
\description{
Plot concordance indices
}
