% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModel.R
\name{create_msm}
\alias{create_msm}
\title{Create a model}
\usage{
create_msm(matrix, states, compile = TRUE)
}
\arguments{
\item{matrix}{a binary matrix where \code{matrix[i,j]} is 1 if transition
from state \code{i} to \code{j} is possible}

\item{states}{a character vector of state names}

\item{compile}{compile 'Stan' model?}
}
\value{
A \code{\link{MultistateModel}} object
}
\description{
Create a model
}
