% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{create_scoring_df}
\alias{create_scoring_df}
\title{Create a data frame to evaluate state visit probability predictions}
\usage{
create_scoring_df(paths_obs, paths_pred, state_name)
}
\arguments{
\item{paths_obs}{A \code{\link{PathData}} object of observed paths.}

\item{paths_pred}{A \code{\link{PathData}} object of predicted paths.}

\item{state_name}{Name of the state of interest.}

\item{df}{A \code{data.frame} with \code{surv} and \code{risk} columns}
}
\description{
Create a data frame to evaluate state visit probability predictions
}
