% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-data.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Creating Stan data list}
\usage{
create_stan_data(model, data)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}} object.}

\item{data}{A \code{\link{JointData}} or \code{\link{PathData}} object.}
}
\value{
A list of data for Stan.
}
\description{
Creating Stan data list
}
\seealso{
Other Stan-related functions: 
\code{\link{create_stan_model}()},
\code{\link{default_stan_filepath}()},
\code{\link{ensure_exposed_stan_functions}()},
\code{\link{fit_stan}()},
\code{\link{plot_stan_data_integral}()},
\code{\link{plot_stan_data_matrix}()}
}
\concept{Stan-related functions}
