% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{create_stan_data_model}
\alias{create_stan_data_model}
\title{Creating Stan data list (model properties)}
\usage{
create_stan_data_model(model)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}} object.}
}
\value{
A list of data for Stan.
}
\description{
Creating Stan data list (model properties)
}
