% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{create_stan_model}
\alias{create_stan_model}
\title{Create the main 'Stan' model}
\usage{
create_stan_model(...)
}
\arguments{
\item{...}{Arguments passed to 'CmdStanR'}
}
\description{
Create the main 'Stan' model
}
