% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{create_standata}
\alias{create_standata}
\title{Create the 'Stan' data list.}
\usage{
create_standata(model, pd, prior_only = FALSE)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}} object.}

\item{pd}{A \code{\link{PathData}} object of observed paths.}

\item{prior_only}{Sample from prior only.}
}
\value{
A list.
}
\description{
Create the 'Stan' data list.
}
