% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{default_stan_filepath}
\alias{default_stan_filepath}
\title{Get path to default Stan file}
\usage{
default_stan_filepath()
}
\description{
Get path to default Stan file
}
\seealso{
Other Stan-related functions: 
\code{\link{create_stan_model}()},
\code{\link{ensure_exposed_stan_functions}()},
\code{\link{fit_stan}()},
\code{\link{plot_stan_data_integral}()},
\code{\link{plot_stan_data_matrix}()}
}
\concept{Stan-related functions}
