% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{do_split}
\alias{do_split}
\title{Train-test split for a transitions data frame}
\usage{
do_split(dt)
}
\arguments{
\item{dt}{A transitions data frame}
}
\value{
a list with test and train subject indices
}
\description{
Train-test split for a transitions data frame
}
