% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DosingData.R
\name{DosingData}
\alias{DosingData}
\title{Dosing data class (R6 class)}
\description{
Dosing data class (R6 class)

Dosing data class (R6 class)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subject_ids}}{Subject ids.
Initialize}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DosingData-set_sub_ids}{\code{DosingData$set_sub_ids()}}
\item \href{#method-DosingData-print}{\code{DosingData$print()}}
\item \href{#method-DosingData-num_subjects}{\code{DosingData$num_subjects()}}
\item \href{#method-DosingData-clone}{\code{DosingData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DosingData-set_sub_ids"></a>}}
\if{latex}{\out{\hypertarget{method-DosingData-set_sub_ids}{}}}
\subsection{Method \code{set_sub_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DosingData$set_sub_ids(subject_ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject_ids}}{A character vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DosingData-print"></a>}}
\if{latex}{\out{\hypertarget{method-DosingData-print}{}}}
\subsection{Method \code{print()}}{
Print info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DosingData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DosingData-num_subjects"></a>}}
\if{latex}{\out{\hypertarget{method-DosingData-num_subjects}{}}}
\subsection{Method \code{num_subjects()}}{
Get number of subjects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DosingData$num_subjects()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DosingData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DosingData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DosingData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
