% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{.brier_score2}
\alias{.brier_score2}
\title{More efficient version of SurvMetrics::Brier}
\usage{
.brier_score2(object, pre_sp, t_star, gt_times = NULL, gt_tstar = NULL)
}
\description{
More efficient version of SurvMetrics::Brier
}
