% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{estimate_average_hazard}
\alias{estimate_average_hazard}
\title{Estimate average hazard of an 'msfit'}
\usage{
estimate_average_hazard(msfit)
}
\arguments{
\item{msfit}{An \code{msfit} object}
}
\description{
Estimate average hazard of an 'msfit'
}
