% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{event_risk}
\alias{event_risk}
\title{Compute risk of event for each subject based on generated paths}
\usage{
event_risk(p, event)
}
\arguments{
\item{p}{A \code{\link{PathData}} object of generated paths}

\item{event}{Name of the event of interest (character)}
}
\value{
A \code{\link{PathData}} object
}
\description{
Compute risk of event for each subject based on generated paths
}
