% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{example_sim_setup_illnessdeath}
\alias{example_sim_setup_illnessdeath}
\title{Example simulation setup}
\usage{
example_sim_setup_illnessdeath(beta_dose = -0.5, beta_age = 0.5)
}
\arguments{
\item{beta_dose}{effect of dose on illness rate}

\item{beta_age}{effect of age on death rate}
}
\value{
a list containing a model and true betas
}
\description{
Example simulation setup
}
