% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{fit_coxph}
\alias{fit_coxph}
\title{Fit Cox PH model using Breslow method}
\usage{
fit_coxph(msdat, formula_rhs = NULL)
}
\arguments{
\item{msdat}{\code{msdata} object}

\item{formula_rhs}{Formula right hand side that is appended to
\code{Surv(Tstart, Tstop, status) ~ }. If \code{NULL} (default), then
\code{strata(trans)} is used}
}
\value{
value returned by \code{survival::coxph}
}
\description{
Fit Cox PH model using Breslow method
}
