% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{fit_mstate}
\alias{fit_mstate}
\title{Fit 'mstate' model}
\usage{
fit_mstate(msdat, formula = NULL)
}
\arguments{
\item{msdat}{\code{msdata} object}

\item{formula}{formula}
}
\description{
Fit 'mstate' model
}
