% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathgen.R
\name{generate_paths_many_subjects}
\alias{generate_paths_many_subjects}
\title{Generate all paths for some subjects}
\usage{
generate_paths_many_subjects(
  fit,
  sd,
  covs,
  TFI,
  init_state,
  t_start,
  t_max_gen,
  state_names,
  trans_names,
  terminal_states,
  df_subjects,
  num_paths = NULL,
  oos = FALSE,
  discretize = TRUE,
  n_repeats = 1,
  use_future = TRUE,
  use_precomp = TRUE,
  tol = 1.03
)
}
\arguments{
\item{fit}{Stan model fit}

\item{sd}{Stan data}

\item{covs}{covariates}

\item{TFI}{the TFI matrix}

\item{init_state}{state where to initialize the paths}

\item{t_start}{start time}

\item{t_max_gen}{max time}

\item{state_names}{state names}

\item{trans_names}{transition names}

\item{terminal_states}{names of terminal states}

\item{df_subjects}{subjects data frame}

\item{num_paths}{number of draws for which to generate paths}

\item{oos}{out-of-sample mode?}

\item{discretize}{discretize the times?}

\item{n_repeats}{number of times to repeat path generation for each draw}

\item{use_future}{use the \code{future} package?}

\item{use_precomp}{use precomputation?}

\item{tol}{multiplier for upper bound of hazard}
}
\value{
TODO: this function could be written more compactly
with less arguments
}
\description{
Generate all paths for some subjects
}
