% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{get_and_format_log_C_haz_draws}
\alias{get_and_format_log_C_haz_draws}
\title{Extract log hazard multiplier draws}
\usage{
get_and_format_log_C_haz_draws(fit, oos)
}
\arguments{
\item{fit}{fit object}

\item{oos}{for out-of-sample subjects?}
}
\value{
an array with \code{dim = c(n_draws, n_hazards, n_obs)}
}
\description{
Extract log hazard multiplier draws
}
