% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{get_and_format_log_h0_draws}
\alias{get_and_format_log_h0_draws}
\title{Extract log baseline hazard draws}
\usage{
get_and_format_log_h0_draws(fit)
}
\arguments{
\item{fit}{fit object}
}
\value{
an array with \code{dim = c(n_draws, n_hazards, n_pred)}
}
\description{
Extract log baseline hazard draws
}
