% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{get_inst_hazard_param_draws}
\alias{get_inst_hazard_param_draws}
\title{Extract and reshape draws of instant hazard related parameters}
\usage{
get_inst_hazard_param_draws(fit)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelStanFit}} object}
}
\value{
a list with elements \code{log_m}, \code{log_w0}, \code{w}, each
of which is an array where the first dimension is number of subjects times
number of draws
}
\description{
Extract and reshape draws of instant hazard related parameters
}
