% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelStanFit.R
\name{hazard_multipliers}
\alias{hazard_multipliers}
\title{Compute hazard multipliers}
\usage{
hazard_multipliers(fit, data = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelStanFit}} object}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used.}
}
\description{
Compute hazard multipliers
}
