% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{illnessdeath_transmat}
\alias{illnessdeath_transmat}
\title{Illness-death model transition matrix}
\usage{
illnessdeath_transmat(state_names = c("Randomization", "Illness", "Death"))
}
\value{
A \code{\link{TransitionMatrix}}
}
\description{
Illness-death model transition matrix
}
