% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointData.R
\name{JointData}
\alias{JointData}
\title{Joint data class (R6 class)}
\description{
Joint data class (R6 class)

Joint data class (R6 class)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{paths}}{The events data, an object of class \code{\link{PathData}}.}

\item{\code{dosing}}{The dosing data, an object of class \code{\link{DosingData}}.
Initialize}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JointData-new}{\code{JointData$new()}}
\item \href{#method-JointData-filter_subjects}{\code{JointData$filter_subjects()}}
\item \href{#method-JointData-print}{\code{JointData$print()}}
\item \href{#method-JointData-clone}{\code{JointData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointData-new"></a>}}
\if{latex}{\out{\hypertarget{method-JointData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointData$new(paths, dosing)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paths}}{The events data, an object of class \code{\link{PathData}}.}

\item{\code{dosing}}{The dosing data, an object of class \code{\link{DosingData}}.
Can be also \code{NULL}.
Filter subjects, creates new object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointData-filter_subjects"></a>}}
\if{latex}{\out{\hypertarget{method-JointData-filter_subjects}{}}}
\subsection{Method \code{filter_subjects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointData$filter_subjects(subject_ids_keep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject_ids_keep}}{Subjects to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointData-print"></a>}}
\if{latex}{\out{\hypertarget{method-JointData-print}{}}}
\subsection{Method \code{print()}}{
Print info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JointData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
