% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{legend_to_PT_matrix}
\alias{legend_to_PT_matrix}
\title{Create matrix indicating possible transitions given current state}
\usage{
legend_to_PT_matrix(legend)
}
\arguments{
\item{legend}{PathData legend}
}
\value{
a binary matrix
}
\description{
Create matrix indicating possible transitions given current state
}
