% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{legend_to_TFI_matrix}
\alias{legend_to_TFI_matrix}
\title{State/transition legend to TFI matrix}
\usage{
legend_to_TFI_matrix(legend)
}
\arguments{
\item{legend}{PathData legend}
}
\value{
an integer matrix
}
\description{
State/transition legend to TFI matrix
}
