% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelStanFit.R
\name{log_hazard_multipliers}
\alias{log_hazard_multipliers}
\title{Compute log_hazard multipliers}
\usage{
log_hazard_multipliers(fit, data = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelStanFit}} object}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used.}
}
\description{
Compute log_hazard multipliers
}
