% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{msfit_plot_cumhaz}
\alias{msfit_plot_cumhaz}
\title{Plot cumulative hazard of 'msfit'}
\usage{
msfit_plot_cumhaz(msfit, legend = NULL)
}
\arguments{
\item{msfit}{An \code{msfit} object}

\item{legend}{transition name legend}
}
\description{
Plot cumulative hazard of 'msfit'
}
