% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{msmfit_check_xpsr_norm}
\alias{msmfit_check_xpsr_norm}
\title{Check exposure normalization}
\usage{
msmfit_check_xpsr_norm(fit)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}
}
\description{
Check exposure normalization
}
