% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{msmfit_exposure}
\alias{msmfit_exposure}
\title{Compute exposure}
\usage{
msmfit_exposure(fit, data = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used. If not \code{NULL}, out-of-sample
mode is assumed (new subjects).}
}
\description{
Compute exposure
}
