% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{msmfit_inst_hazard_param_draws}
\alias{msmfit_inst_hazard_param_draws}
\title{Extract and reshape draws of instant hazard related parameters}
\usage{
msmfit_inst_hazard_param_draws(fit, oos = FALSE, data = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{oos}{Out-of-sample mode? If \code{FALSE}, the possible subject-specific
fitted parameters are used. If \code{TRUE}, acting
as if the subjects are new.}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used.}
}
\value{
a list with elements \code{log_m}, \code{log_w0}, \code{w}, each
of which is an array where the first dimension is the number of subjects times
the number of draws
}
\description{
Extract and reshape draws of instant hazard related parameters
}
