% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{msmfit_log_hazard_multipliers}
\alias{msmfit_log_hazard_multipliers}
\title{Compute log_hazard multipliers}
\usage{
msmfit_log_hazard_multipliers(fit, oos, data = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{oos}{Out-of-sample mode? If \code{FALSE}, the \code{data} input
should be \code{NULL}, and in this case possible, subject specific
fitted parameters are used. If \code{TRUE}, acting
as if the subjects are new.}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used.}
}
\value{
A list of length \code{n_draws} where each element is a
matrix of shape \code{n_subject} x \code{n_transitions}
}
\description{
Compute log_hazard multipliers
}
