% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{msmfit_pk_params}
\alias{msmfit_pk_params}
\title{Evaluate PK parameters}
\usage{
msmfit_pk_params(fit, data = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used. If not \code{NULL}, out-of-sample
mode is assumed (new subjects).}
}
\value{
A list with length equal to number of draws.
}
\description{
Evaluate PK parameters
}
