% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelStanFit.R
\name{msmsf_inst_hazard_param_draws}
\alias{msmsf_inst_hazard_param_draws}
\title{Extract and reshape draws of instant hazard related parameters}
\usage{
msmsf_inst_hazard_param_draws(fit, data = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelStanFit}} object}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used. If not \code{NULL}, out-of-sample
mode is assumed (new subjects).}
}
\value{
a list with elements \code{log_m}, \code{log_w0}, \code{w}, each
of which is an array where the first dimension is number of subjects times
number of draws
}
\description{
Extract and reshape draws of instant hazard related parameters
}
