% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelStanFit.R
\name{msmsf_log_hazard_multipliers}
\alias{msmsf_log_hazard_multipliers}
\title{Compute log_hazard multipliers}
\usage{
msmsf_log_hazard_multipliers(fit, data = NULL)
}
\description{
Compute log_hazard multipliers
}
