% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModel.R
\name{MultiStateModel}
\alias{MultiStateModel}
\title{Main model class}
\description{
Main model class

Main model class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiStateModel-new}{\code{MultiStateModel$new()}}
\item \href{#method-MultiStateModel-get_stan_model}{\code{MultiStateModel$get_stan_model()}}
\item \href{#method-MultiStateModel-create_MultiStateModel}{\code{MultiStateModel$create_MultiStateModel()}}
\item \href{#method-MultiStateModel-compile}{\code{MultiStateModel$compile()}}
\item \href{#method-MultiStateModel-clone}{\code{MultiStateModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiStateModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiStateModel-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStateModel$new(compile = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{compile}}{Should the 'Stan' model code be created and compiled.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiStateModel-get_stan_model"></a>}}
\if{latex}{\out{\hypertarget{method-MultiStateModel-get_stan_model}{}}}
\subsection{Method \code{get_stan_model()}}{
Get the underlying 'Stan' model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStateModel$get_stan_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiStateModel-create_MultiStateModel"></a>}}
\if{latex}{\out{\hypertarget{method-MultiStateModel-create_MultiStateModel}{}}}
\subsection{Method \code{create_MultiStateModel()}}{
Create and compile the 'Stan' model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStateModel$create_MultiStateModel(dir = tempdir())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{Path to directory where to store the \code{.stan} file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'CmdStanR' model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiStateModel-compile"></a>}}
\if{latex}{\out{\hypertarget{method-MultiStateModel-compile}{}}}
\subsection{Method \code{compile()}}{
Create and compile the 'Stan' model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStateModel$compile(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \code{create_MultiStateModel()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated model object (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiStateModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiStateModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiStateModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
