% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelStanFit.R
\name{MultistateModelStanFit}
\alias{MultistateModelStanFit}
\title{The Fit class}
\description{
The Fit class

The Fit class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{stan_fit}}{The 'Stan' fit object}

\item{\code{stan_data}}{Full 'Stan' data list}

\item{\code{data}}{A \code{\link{JointData}} object}

\item{\code{model}}{A \code{\link{MultistateModel}} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultistateModelStanFit-get_model}{\code{MultistateModelStanFit$get_model()}}
\item \href{#method-MultistateModelStanFit-new}{\code{MultistateModelStanFit$new()}}
\item \href{#method-MultistateModelStanFit-draws}{\code{MultistateModelStanFit$draws()}}
\item \href{#method-MultistateModelStanFit-print}{\code{MultistateModelStanFit$print()}}
\item \href{#method-MultistateModelStanFit-plot_basisfun}{\code{MultistateModelStanFit$plot_basisfun()}}
\item \href{#method-MultistateModelStanFit-log_z_pars}{\code{MultistateModelStanFit$log_z_pars()}}
\item \href{#method-MultistateModelStanFit-gq}{\code{MultistateModelStanFit$gq()}}
\item \href{#method-MultistateModelStanFit-clone}{\code{MultistateModelStanFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-get_model}{}}}
\subsection{Method \code{get_model()}}{
Get model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$get_model(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of model. Only has effect for
\code{\link{JointModelFit}} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-new}{}}}
\subsection{Method \code{new()}}{
Create model fit object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$new(model, data, stan_fit, stan_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A \code{\link{MultistateModel}} object}

\item{\code{data}}{A \code{\link{JointData}} object}

\item{\code{stan_fit}}{The 'Stan' fit object}

\item{\code{stan_data}}{Full 'Stan' data list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-draws"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-draws}{}}}
\subsection{Method \code{draws()}}{
Extract draws as \code{rvar}s
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$draws(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Param/quantity name
Print the object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-plot_basisfun"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-plot_basisfun}{}}}
\subsection{Method \code{plot_basisfun()}}{
Plot used basis functions (grid)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$plot_basisfun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-log_z_pars"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-log_z_pars}{}}}
\subsection{Method \code{log_z_pars()}}{
Full names of parameters that start with \code{log_z_}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$log_z_pars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-gq"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-gq}{}}}
\subsection{Method \code{gq()}}{
Generate quantities using the fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$gq(stan_data = NULL, fitted_params = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stan_data}}{Full 'Stan' input list.}

\item{\code{fitted_params}}{Argument to \code{generate_quantities()}.}

\item{\code{...}}{Other arguments to \code{generate_quantities()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
