% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{p_state_occupancy}
\alias{p_state_occupancy}
\title{Solve state occupancy probabilities for each subject and draw in
'MultistateModelFit'}
\usage{
p_state_occupancy(
  fit,
  oos = FALSE,
  t_start = 0,
  t_out = NULL,
  data = NULL,
  ...
)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{oos}{Out-of-sample mode? If \code{FALSE}, the \code{data} input
should be \code{NULL}, and in this case possible, subject specific
fitted parameters are used. If \code{TRUE}, acting
as if the subjects are new.}

\item{t_start}{Initial time}

\item{t_out}{End times (vector)}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used.}

\item{...}{Arguments passed to \code{deSolve::ode()}.}
}
\value{
A list with
\itemize{
  \item A 4-dimensional \code{rvar} array \code{P} where \code{P[n,k,,]} is the
    transition matrix for subject \code{n} at the \code{k}th output time
    point
   \item Subject ids in the same order as in the first dimension of \code{P}
   \item The numeric vector \code{t_out}
 }
}
\description{
Solve state occupancy probabilities for each subject and draw in
'MultistateModelFit'
}
