% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{p_state_visit}
\alias{p_state_visit}
\title{For each state, compute probability of visiting it at least once before given time}
\usage{
p_state_visit(pd, t = NULL, by = NULL)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object.}

\item{t}{The given time. If \code{NULL}, is set to \code{max(pd$get_path_df()$time)}.}

\item{by}{Factor to summarize over.}
}
\value{
A data frame
}
\description{
For each state, compute probability of visiting it at least once before given time
}
