% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{p_state_visit_per_subject}
\alias{p_state_visit_per_subject}
\title{For each subject, compute probability of visiting a given state
at least once before given time}
\usage{
p_state_visit_per_subject(pd, state_name, t = NULL)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object.}

\item{state_name}{Name of the state (character).}

\item{t}{The given time. If \code{NULL}, is set to \code{max(pd$get_path_df()$time)}.}
}
\value{
A data frame
}
\description{
Convenient wrapper for \code{\link{p_state_visit}}.
}
