% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{PathData}
\alias{PathData}
\title{Path data class (R6 class)}
\description{
It is not recommended for users to try to create data using
the constructor of this class. Rather use \code{\link{df_to_pathdata}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subject_df}}{Data frame with one row per subject. Must have one
row for each subject, and
\code{subject_id} and all covariates as columns.}

\item{\code{path_df}}{Data frame of actual paths. Each row corresponds to
one time point. Must have \code{path_id},
\code{state}, \code{time}, and \code{trans_idx} as columns. These are
\itemize{
  \item \code{path_id}: path identifier
  \item \code{state}: integer index of the state at the time point, and
  until the next time point
  \item \code{time}: time point
  \item \code{trans_idx}: Integer indicating which transition this
  time point corresponds to. If the time point is not a state transition,
  this should be 0.
}}

\item{\code{link_df}}{Links the path and subject data frames. Must have
one row for each path, and \code{path_id}, \code{draw_idx},
\code{rep_idx}, and \code{subject_id} as columns.}

\item{\code{covs}}{Covariate column names.}

\item{\code{transmat}}{A \code{\link{TransitionMatrix}} describing the system to
which the paths belong.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PathData-new}{\code{PathData$new()}}
\item \href{#method-PathData-unique_subjects}{\code{PathData$unique_subjects()}}
\item \href{#method-PathData-covariate_names}{\code{PathData$covariate_names()}}
\item \href{#method-PathData-subset_covariates}{\code{PathData$subset_covariates()}}
\item \href{#method-PathData-state_at}{\code{PathData$state_at()}}
\item \href{#method-PathData-transition_times}{\code{PathData$transition_times()}}
\item \href{#method-PathData-n_paths}{\code{PathData$n_paths()}}
\item \href{#method-PathData-state_names}{\code{PathData$state_names()}}
\item \href{#method-PathData-get_event_state_names}{\code{PathData$get_event_state_names()}}
\item \href{#method-PathData-terminal_states}{\code{PathData$terminal_states()}}
\item \href{#method-PathData-print}{\code{PathData$print()}}
\item \href{#method-PathData-get_path_df}{\code{PathData$get_path_df()}}
\item \href{#method-PathData-as_data_frame}{\code{PathData$as_data_frame()}}
\item \href{#method-PathData-full_link}{\code{PathData$full_link()}}
\item \href{#method-PathData-as_transitions}{\code{PathData$as_transitions()}}
\item \href{#method-PathData-as_transitions_alt}{\code{PathData$as_transitions_alt()}}
\item \href{#method-PathData-as_msdata}{\code{PathData$as_msdata()}}
\item \href{#method-PathData-plot_paths}{\code{PathData$plot_paths()}}
\item \href{#method-PathData-prop_matrix}{\code{PathData$prop_matrix()}}
\item \href{#method-PathData-plot_graph}{\code{PathData$plot_graph()}}
\item \href{#method-PathData-fit_coxph}{\code{PathData$fit_coxph()}}
\item \href{#method-PathData-fit_mstate}{\code{PathData$fit_mstate()}}
\item \href{#method-PathData-filter}{\code{PathData$filter()}}
\item \href{#method-PathData-clone}{\code{PathData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-new"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-new}{}}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$new(subject_df, path_df, link_df, transmat, covs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject_df}}{Data frame with one row per subject. Must have one
row for each subject, and
\code{subject_id} and all covariates as columns.}

\item{\code{path_df}}{Data frame of actual paths. Each row corresponds to
one time point. Must have \code{path_id},
\code{state}, \code{time}, and \code{trans_idx} as columns. These are
\itemize{
  \item \code{path_id}: path identifier
  \item \code{state}: integer index of the state at the time point, and
  until the next time point
  \item \code{time}: time point
  \item \code{trans_idx}: Integer indicating which transition this
  time point corresponds to. If the time point is not a state transition,
  this should be 0.
}}

\item{\code{link_df}}{Links the path and subject data frames. Must have
one row for each path, and \code{path_id}, \code{draw_idx},
\code{rep_idx}, and \code{subject_id} as columns.}

\item{\code{transmat}}{A \code{\link{TransitionMatrix}} describing the system to
which the paths belong.}

\item{\code{covs}}{Covariate column names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-unique_subjects"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-unique_subjects}{}}}
\subsection{Method \code{unique_subjects()}}{
Get unique subject ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$unique_subjects()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-covariate_names"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-covariate_names}{}}}
\subsection{Method \code{covariate_names()}}{
Get names of covariates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$covariate_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-subset_covariates"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-subset_covariates}{}}}
\subsection{Method \code{subset_covariates()}}{
Create a new PathData with a subset of covariates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$subset_covariates(covs, renamed_old = NULL, renamed_new = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covs}}{Names of all new covariates}

\item{\code{renamed_old}}{Name of an old covariate to rename}

\item{\code{renamed_new}}{Name of new covariate where old one is copied}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-state_at"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-state_at}{}}}
\subsection{Method \code{state_at()}}{
For each path, get the state it is in at time t
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$state_at(t)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{time}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame with one row for each path
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-transition_times"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-transition_times}{}}}
\subsection{Method \code{transition_times()}}{
Get all transition times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$transition_times(trans_inds = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trans_inds}}{Indices of transition whose occurrence times are
requested. Default is \code{NULL}, in which case all transitions are
considered.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-n_paths"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-n_paths}{}}}
\subsection{Method \code{n_paths()}}{
Get number of paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$n_paths()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-state_names"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-state_names}{}}}
\subsection{Method \code{state_names()}}{
Get names of all states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$state_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-get_event_state_names"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-get_event_state_names}{}}}
\subsection{Method \code{get_event_state_names()}}{
Get names of all states to which transitioning
can be a transition event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$get_event_state_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-terminal_states"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-terminal_states}{}}}
\subsection{Method \code{terminal_states()}}{
Get names of terminal states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$terminal_states()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-print"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-print}{}}}
\subsection{Method \code{print()}}{
Print info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-get_path_df"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-get_path_df}{}}}
\subsection{Method \code{get_path_df()}}{
Get path data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$get_path_df(truncate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{truncate}}{Remove rows after terminal events?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_data_frame}{}}}
\subsection{Method \code{as_data_frame()}}{
Convert to one long data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_data_frame(covariates = NULL, truncate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Which covariates to include?}

\item{\code{truncate}}{Remove rows after terminal events?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-full_link"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-full_link}{}}}
\subsection{Method \code{full_link()}}{
Full link data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$full_link(covariates = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Which covariates to include}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with same number of rows as \code{link_df},
including also the covariate columns and \code{subject_id}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_transitions"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_transitions}{}}}
\subsection{Method \code{as_transitions()}}{
Data frame in transitions format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_transitions(covariates = NULL, truncate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Which covariates to include?}

\item{\code{truncate}}{Remove rows after terminal events first?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_transitions_alt"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_transitions_alt}{}}}
\subsection{Method \code{as_transitions_alt()}}{
Data frame in alternative transitions format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_transitions_alt(covariates = NULL, truncate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Which covariates to include?}

\item{\code{truncate}}{Remove rows after terminal events first?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_msdata"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_msdata}{}}}
\subsection{Method \code{as_msdata()}}{
Convert to format used by the 'mstate' package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_msdata(covariates = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Which covariates to include?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{msdata} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-plot_paths"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-plot_paths}{}}}
\subsection{Method \code{plot_paths()}}{
Step plot of the paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$plot_paths(n_paths = NULL, alpha = 0.5, truncate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_paths}}{Number of paths to subsample for plotting.}

\item{\code{alpha}}{opacity}

\item{\code{truncate}}{truncate after terminal events?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-prop_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-prop_matrix}{}}}
\subsection{Method \code{prop_matrix()}}{
Transition proportion matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$prop_matrix(include_censor = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_censor}}{Include censoring (no event) proportion in the
matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{table}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-plot_graph"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-plot_graph}{}}}
\subsection{Method \code{plot_graph()}}{
Visualize the transition proportion matrix as a graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$plot_graph(digits = 3, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{Max number of digits to show in numbers}

\item{\code{...}}{Arguments passed to \code{qgraph}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{qgraph} plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-fit_coxph"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-fit_coxph}{}}}
\subsection{Method \code{fit_coxph()}}{
Fit Cox proportional hazards model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$fit_coxph(covariates = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Covariates to include.}

\item{\code{...}}{Arguments passed to \code{survival::coxph}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-fit_mstate"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-fit_mstate}{}}}
\subsection{Method \code{fit_mstate()}}{
Fit frequentist 'mstate' model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$fit_mstate(covariates = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Covariates to include.}

\item{\code{...}}{Arguments passed to \code{survival::coxph}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-filter"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-filter}{}}}
\subsection{Method \code{filter()}}{
Filter based on subject id, creates new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$filter(subject_ids_keep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject_ids_keep}}{Subject ids to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
