% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DosingData.R
\name{pk_2cpt_pss}
\alias{pk_2cpt_pss}
\title{Partially steady-state PK model}
\usage{
pk_2cpt_pss(t, dose_ss, times, doses, theta, tau)
}
\arguments{
\item{t}{vector of output time points}

\item{dose_ss}{dose amount in SS (for each subject)}

\item{times}{time points, first of which is the end of stedy-state assumption}

\item{doses}{doses taken after \code{t_last_ss}}

\item{theta}{PK params for each subject}

\item{tau}{Dosing interval (same for all subjects).}
}
\value{
For each subject, the concentration in the central compartment at
times \code{t}
}
\description{
For each subject
}
