% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKModel.R
\name{PKModel}
\alias{PKModel}
\title{One-compartment PK Model}
\description{
One-compartment PK Model

One-compartment PK Model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PKModel-new}{\code{PKModel$new()}}
\item \href{#method-PKModel-ka_covs}{\code{PKModel$ka_covs()}}
\item \href{#method-PKModel-CL_covs}{\code{PKModel$CL_covs()}}
\item \href{#method-PKModel-V2_covs}{\code{PKModel$V2_covs()}}
\item \href{#method-PKModel-print}{\code{PKModel$print()}}
\item \href{#method-PKModel-simulate_ss}{\code{PKModel$simulate_ss()}}
\item \href{#method-PKModel-clone}{\code{PKModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$new(covariates)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{A list with elements \code{ka},
\code{CL}, and \code{V2}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-ka_covs"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-ka_covs}{}}}
\subsection{Method \code{ka_covs()}}{
Get the covariates affecting the k_a parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$ka_covs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-CL_covs"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-CL_covs}{}}}
\subsection{Method \code{CL_covs()}}{
Get the covariates affecting the CL parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$CL_covs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-V2_covs"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-V2_covs}{}}}
\subsection{Method \code{V2_covs()}}{
Get the covariates affecting the V_2 parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$V2_covs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-print}{}}}
\subsection{Method \code{print()}}{
Print the object info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-simulate_ss"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-simulate_ss}{}}}
\subsection{Method \code{simulate_ss()}}{
Simulate system in steady state
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$simulate_ss(t, theta, dose, tau)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{output time points}

\item{\code{theta}}{parameter values}

\item{\code{dose}}{dose}

\item{\code{tau}}{dosing interval}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector with the same length as \code{t}, representing the
concentration in the central compartment.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
