% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{pkmstate-package}
\alias{pkmstate-package}
\alias{pkmstate}
\title{The 'pkmstate' package.}
\description{
Bayesian multistate modeling using 'Stan'.
}
\section{Getting started}{

See the following \code{R6} classes.
\itemize{
 \item \code{\link{PathData}}: Class to hold generated paths.
}
}

\author{
Juho Timonen (first.last at iki.fi) and Jacki Buros
}
\keyword{Bayesian}
\keyword{Stan}
\keyword{multistate}
