% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{plot_cumhaz_msfit}
\alias{plot_cumhaz_msfit}
\title{Plot cumulative hazard of 'msfit'}
\usage{
plot_cumhaz_msfit(msfit, legend = NULL)
}
\arguments{
\item{msfit}{An \code{msfit} object}

\item{legend}{transition name legend}
}
\description{
Plot cumulative hazard of 'msfit'
}
