% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_h0}
\alias{plot_h0}
\title{Plot log baseline hazard}
\usage{
plot_h0(fit, sd, pd, log_h0_true = NULL)
}
\arguments{
\item{fit}{fit object}

\item{sd}{Stan data list}

\item{pd}{A \code{\link{PathData}} object used to create the Stan data.}

\item{log_h0_true}{true baseline hazards}
}
\description{
Plot log baseline hazard
}
