% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_other_beta}
\alias{plot_other_beta}
\title{Plot coefficients for other covariates}
\usage{
plot_other_beta(fit, stan_dat, pd, df_beta_true = NULL)
}
\arguments{
\item{fit}{fit object}

\item{stan_dat}{stan data list}

\item{pd}{A \code{\link{PathData}} object used to create the Stan data.}

\item{df_beta_true}{must be a data frame with columns \code{trans_idx},
\code{cov_name} and \code{beta_true}}
}
\description{
Plot coefficients for other covariates
}
