% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_p_event}
\alias{plot_p_event}
\title{Plot event rates over all subjects}
\usage{
plot_p_event(pd, paths_gen, paths_gen_oos, names)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object of observed data}

\item{paths_gen}{A \code{\link{PathData}} object of generated paths
(in-sample)}

\item{paths_gen_oos}{A \code{\link{PathData}} object of generated paths
(out-of-sample)}

\item{names}{names of the three former objects}
}
\description{
Plot event rates over all subjects
}
