% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_p_event_by_dose}
\alias{plot_p_event_by_dose}
\title{Plot event rates over all subjects by dose}
\usage{
plot_p_event_by_dose(pd, paths_gen, paths_gen_oos, names, t_max = 365.25 * 3)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object of observed data}

\item{paths_gen}{A \code{\link{PathData}} object of generated paths
(in-sample)}

\item{paths_gen_oos}{A \code{\link{PathData}} object of generated paths
(out-of-sample)}

\item{names}{names of the three former objects}

\item{t_max}{max time used in path generation}
}
\description{
Plot event rates over all subjects by dose
}
