% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-plot.R
\name{plot_stan_data}
\alias{plot_stan_data}
\title{Visualize parts of the 'Stan' data}
\usage{
plot_stan_data(model, stan_data, name, subject_idx)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}}.}

\item{stan_data}{A list of data for Stan. Created using
\code{\link{create_stan_data}}.}

\item{name}{Name of Stan data field to plot. Currently must be one of
\code{"transition", "at_risk"}.}

\item{subject_idx}{Index of subject whose data rows to plot.}
}
\description{
Visualize parts of the 'Stan' data
}
