% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-plot.R
\name{plot_stan_data_integral}
\alias{plot_stan_data_integral}
\title{Visualize numerical integration of hazard}
\usage{
plot_stan_data_integral(model, stan_data, iidx, h0 = 0.001, w = NULL)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}}.}

\item{stan_data}{A list of data for Stan. Created using
\code{\link{create_stan_data}}.}

\item{iidx}{Interval index}

\item{h0}{Hazard function average value}

\item{w}{Hazard function spline weights. Randomly generated if not given.}

\item{subject_idx}{Index of subject whose data rows to plot.}
}
\description{
Visualize numerical integration of hazard
}
\seealso{
Other Stan-related functions: 
\code{\link{create_stan_model}()},
\code{\link{default_stan_filepath}()},
\code{\link{ensure_exposed_stan_functions}()},
\code{\link{fit_stan}()},
\code{\link{plot_stan_data_matrix}()}
}
\concept{Stan-related functions}
