% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{plot_state_occupancy}
\alias{plot_state_occupancy}
\title{Plot mean state occupancy probabilities over time for each subject}
\usage{
plot_state_occupancy(df)
}
\arguments{
\item{df}{Data frame returned by \code{\link{p_state_occupancy}}}
}
\description{
Plot mean state occupancy probabilities over time for each subject
}
