% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk.R
\name{pop_2cpt_partly_ss}
\alias{pop_2cpt_partly_ss}
\title{Partially steady-state PK model}
\usage{
pop_2cpt_partly_ss(t, dose_ss, times, doses, theta, tau)
}
\arguments{
\item{t}{vector of output time points}

\item{dose_ss}{dose amount in SS (for each subject)}

\item{times}{time points after \code{t_last_ss}}

\item{doses}{doses taken after \code{t_last_ss}}

\item{theta}{PK params for each subject}

\item{tau}{Dosing interval (same for all subjects).}
}
\description{
For each subject
}
