% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DosingData.R
\name{PSSDosingData}
\alias{PSSDosingData}
\title{Partially steady-state dosing data class (R6 class)}
\description{
Contains information about taken doses and dose times.
}
\section{Super class}{
\code{\link[bmstate:DosingData]{bmstate::DosingData}} -> \code{PSSDosingData}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{doses}}{Dose amounts (list with length equal to number of subjects).
Corresponds to doses after the steady state.}

\item{\code{times}}{Dose times (list with length equal to number of subjects).
Corresponds to doses after the steady state. First time here is the end
of steady-state assumption time range.}

\item{\code{dose_ss}}{Steady-state dose.}

\item{\code{tau_ss}}{Steady-state dosing interval.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PSSDosingData-new}{\code{PSSDosingData$new()}}
\item \href{#method-PSSDosingData-as_data_frame}{\code{PSSDosingData$as_data_frame()}}
\item \href{#method-PSSDosingData-simulate_pk}{\code{PSSDosingData$simulate_pk()}}
\item \href{#method-PSSDosingData-plot}{\code{PSSDosingData$plot()}}
\item \href{#method-PSSDosingData-filter}{\code{PSSDosingData$filter()}}
\item \href{#method-PSSDosingData-clone}{\code{PSSDosingData$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bmstate" data-topic="DosingData" data-id="num_subjects"><a href='../../bmstate/html/DosingData.html#method-DosingData-num_subjects'><code>bmstate::DosingData$num_subjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmstate" data-topic="DosingData" data-id="print"><a href='../../bmstate/html/DosingData.html#method-DosingData-print'><code>bmstate::DosingData$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmstate" data-topic="DosingData" data-id="set_sub_ids"><a href='../../bmstate/html/DosingData.html#method-DosingData-set_sub_ids'><code>bmstate::DosingData$set_sub_ids()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSSDosingData-new"></a>}}
\if{latex}{\out{\hypertarget{method-PSSDosingData-new}{}}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSSDosingData$new(subject_ids, doses, times, dose_ss = NULL, tau_ss = 24)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject_ids}}{A character vector}

\item{\code{doses}}{Dose amounts (list with length equal to number of subjects).
Corresponds to doses after the steady state.}

\item{\code{times}}{Dose times (list with length equal to number of subjects).
Corresponds to doses after the steady state. First time here is the end
of steady-state assumption time range.}

\item{\code{dose_ss}}{Steady-state dose.}

\item{\code{tau_ss}}{Steady-state dosing interval.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSSDosingData-as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-PSSDosingData-as_data_frame}{}}}
\subsection{Method \code{as_data_frame()}}{
As data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSSDosingData$as_data_frame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSSDosingData-simulate_pk"></a>}}
\if{latex}{\out{\hypertarget{method-PSSDosingData-simulate_pk}{}}}
\subsection{Method \code{simulate_pk()}}{
Simulate PK dynamics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSSDosingData$simulate_pk(t, theta, MAX_CONC, skip_assert = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{A vector of output times for each subject (a list).}

\item{\code{theta}}{A matrix of parameters.}

\item{\code{MAX_CONC}}{concentration upper bound}

\item{\code{skip_assert}}{Skip most assertions and call exposed Stan directly,
assuming that it exists?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSSDosingData-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PSSDosingData-plot}{}}}
\subsection{Method \code{plot()}}{
Plot dosing (and PK) data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSSDosingData$plot(
  df_fit = NULL,
  subject_df = NULL,
  max_num_subjects = 12,
  subject_ids = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_fit}}{Fit data frame. Uses columns \code{val}, \code{lower}
and \code{upper}.}

\item{\code{subject_df}}{Subject data frame.}

\item{\code{max_num_subjects}}{Max number of subjects to plot.}

\item{\code{subject_ids}}{Which subjects to plot?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSSDosingData-filter"></a>}}
\if{latex}{\out{\hypertarget{method-PSSDosingData-filter}{}}}
\subsection{Method \code{filter()}}{
Filter based on subject id, creates new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSSDosingData$filter(subject_ids_keep = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject_ids_keep}}{Subject ids to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PSSDosingData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PSSDosingData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PSSDosingData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
