% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simulate_multitransition_data}
\alias{simulate_multitransition_data}
\title{Data simulation (multitransition)}
\usage{
simulate_multitransition_data(
  N_subject,
  covs,
  h0_true,
  df_beta_true,
  sys_idx = 1
)
}
\arguments{
\item{N_subject}{number of subjects}

\item{covs}{covariates}

\item{h0_true}{true baseline hazard value (constant)}

\item{df_beta_true}{true covariate effects}

\item{sys_idx}{index of simulation system}
}
\description{
Data simulation (multitransition)
}
