% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simulate_singletransition_data}
\alias{simulate_singletransition_data}
\title{Data simulation}
\usage{
simulate_singletransition_data(
  N_subject,
  covs,
  h0_true = 0.001,
  df_beta_true = NULL
)
}
\arguments{
\item{N_subject}{number of subjects}

\item{covs}{covariates}

\item{h0_true}{true baseline hazard value (constant)}

\item{df_beta_true}{true covariate effects}
}
\description{
Data simulation
}
