% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelStanFit.R
\name{solve_trans_prob_fit}
\alias{solve_trans_prob_fit}
\title{Solve transition probabilities for each subject in 'MultistateModelStanFit'}
\usage{
solve_trans_prob_fit(fit, init_state = 1, t_init = 0, t_end = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelStanFit}} object}

\item{init_state}{Index of initial state}

\item{t_init}{Initial time}

\item{t_end}{End time}
}
\value{
A data frame, where each row has the probabilities that a given subject
will be in each state at time \code{t_end} given that they were in
\code{init_state} at \code{t_init}
}
\description{
Solve transition probabilities for each subject in 'MultistateModelStanFit'
}
