% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_time_evolution.R
\name{solve_trans_prob_matrix_each_subject}
\alias{solve_trans_prob_matrix_each_subject}
\title{Solve transition probability matrices for each subject and draw in
'MultistateModelFit'}
\usage{
solve_trans_prob_matrix_each_subject(
  fit,
  t_start = 0,
  t_out = NULL,
  data = NULL,
  ...
)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{t_start}{Initial time}

\item{t_out}{End times (vector)}

\item{...}{Arguments passed to \code{deSolve::ode()}.}
}
\value{
A list with
\itemize{
  \item A 4-dimensional \code{rvar} array \code{P} where \code{P[n,k,,]} is the
    transition matrix for subject \code{n} at the \code{k}th output time
    point
   \item Subject ids in the same order as in the first dimension of \code{P}
   \item The numeric vector \code{t_out}
 }
}
\description{
Solve transition probability matrices for each subject and draw in
'MultistateModelFit'
}
