% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{split_data}
\alias{split_data}
\title{Train-test split}
\usage{
split_data(dat, p_test = 0.25)
}
\arguments{
\item{dat}{A \code{\link{JointData}} object}

\item{p_test}{Proportion of test subjects}
}
\value{
a list with test and train \code{\link{JointData}} objects
}
\description{
Train-test split
}
