% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{StanMultistateModelFit}
\alias{StanMultistateModelFit}
\title{The Fit class}
\description{
The Fit class

The Fit class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{stan_fit}}{The Stan fit object}

\item{\code{stan_data}}{Full Stan data list}

\item{\code{data}}{A \code{\link{JointData}} object}

\item{\code{model}}{A \code{\link{MultistateModel}} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanMultistateModelFit-get_model}{\code{StanMultistateModelFit$get_model()}}
\item \href{#method-StanMultistateModelFit-new}{\code{StanMultistateModelFit$new()}}
\item \href{#method-StanMultistateModelFit-draws}{\code{StanMultistateModelFit$draws()}}
\item \href{#method-StanMultistateModelFit-log_z_pars}{\code{StanMultistateModelFit$log_z_pars()}}
\item \href{#method-StanMultistateModelFit-gq}{\code{StanMultistateModelFit$gq()}}
\item \href{#method-StanMultistateModelFit-clone}{\code{StanMultistateModelFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMultistateModelFit-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-StanMultistateModelFit-get_model}{}}}
\subsection{Method \code{get_model()}}{
Get model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMultistateModelFit$get_model(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of model. Only has effect for
\code{\link{JointModelFit}} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMultistateModelFit-new"></a>}}
\if{latex}{\out{\hypertarget{method-StanMultistateModelFit-new}{}}}
\subsection{Method \code{new()}}{
Create model fit object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMultistateModelFit$new(model, data, stan_fit, stan_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A \code{\link{MultistateModel}} object}

\item{\code{data}}{A \code{\link{JointData}} object}

\item{\code{stan_fit}}{The Stan fit object}

\item{\code{stan_data}}{Full Stan data list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMultistateModelFit-draws"></a>}}
\if{latex}{\out{\hypertarget{method-StanMultistateModelFit-draws}{}}}
\subsection{Method \code{draws()}}{
Extract draws as \code{rvar}s
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMultistateModelFit$draws(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Param/quantity name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMultistateModelFit-log_z_pars"></a>}}
\if{latex}{\out{\hypertarget{method-StanMultistateModelFit-log_z_pars}{}}}
\subsection{Method \code{log_z_pars()}}{
Full names of parameters that start with 'log_z_'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMultistateModelFit$log_z_pars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMultistateModelFit-gq"></a>}}
\if{latex}{\out{\hypertarget{method-StanMultistateModelFit-gq}{}}}
\subsection{Method \code{gq()}}{
Generate quantities using the fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMultistateModelFit$gq(stan_data = NULL, fitted_params = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stan_data}}{Full 'Stan' input list.}

\item{\code{fitted_params}}{Argument to \code{generate_quantities()}.}

\item{\code{...}}{Other arguments to \code{generate_quantities()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMultistateModelFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StanMultistateModelFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMultistateModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
