% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{subject_df_with_idx}
\alias{subject_df_with_idx}
\title{Get subject df with numeric subject index}
\usage{
subject_df_with_idx(pd, subs, id_map)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object}

\item{subs}{Subject indices (from \code{\link{do_split}})}

\item{id_map}{Maps numeric id to original id}
}
\description{
Get subject df with numeric subject index
}
