% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{summarize_ppsurv}
\alias{summarize_ppsurv}
\title{Summarize paths using 'ppsurv'}
\usage{
summarize_ppsurv(
  pd,
  target_times,
  by = "subject_id",
  split_by = by,
  truncate_at_terminal_events = TRUE,
  use_old = FALSE
)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object}

\item{target_times}{Times where to summarize}

\item{by}{Factor by which to summarize}

\item{split_by}{Factor by which to split}

\item{truncate_at_terminal_events}{Truncate paths at terminal events?
TODO: explain this better}

\item{use_old}{Use old version?}
}
\description{
Summarize paths using 'ppsurv'
}
