% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{TransitionMatrix}
\alias{TransitionMatrix}
\title{Defines states and possible transitions}
\description{
Defines states and possible transitions

Defines states and possible transitions
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{matrix}}{a binary matrix where \code{matrix[i,j]} is 1 if transition
from state \code{i} to \code{j} is possible}

\item{\code{states}}{a character vector of state names}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransitionMatrix-new}{\code{TransitionMatrix$new()}}
\item \href{#method-TransitionMatrix-print}{\code{TransitionMatrix$print()}}
\item \href{#method-TransitionMatrix-absorbing_states}{\code{TransitionMatrix$absorbing_states()}}
\item \href{#method-TransitionMatrix-source_states}{\code{TransitionMatrix$source_states()}}
\item \href{#method-TransitionMatrix-clone}{\code{TransitionMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-new"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-new}{}}}
\subsection{Method \code{new()}}{
Create graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$new(matrix, states)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{Binary square matrix that defines the states and possible}

\item{\code{states}}{Names of the states
transitions
Print output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-print"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
Get states that cannot be transitioned from
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-absorbing_states"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-absorbing_states}{}}}
\subsection{Method \code{absorbing_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$absorbing_states(names = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Return names of the states? Otherwise returns indices.
Get states that cannot be transitioned to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-source_states"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-source_states}{}}}
\subsection{Method \code{source_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$source_states(names = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Return names of the states? Otherwise returns indices.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
