% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{transmat}
\alias{transmat}
\alias{transmat_full}
\alias{transmat_comprisk}
\alias{transmat_survival}
\alias{transmat_illnessdeath}
\alias{transmat_diamond}
\title{Create a common transition matrix}
\usage{
transmat_full(
  state_names = LETTERS[1:4],
  sources = 1,
  terminal = length(state_names),
  self_loops = TRUE
)

transmat_comprisk(state_names = LETTERS[1:4])

transmat_survival(state_names = c("Alive", "Dead"))

transmat_illnessdeath(state_names = c("Randomization", "Illness", "Death"))

transmat_diamond(state_names = LETTERS[1:4])
}
\arguments{
\item{state_names}{Names of the states. The length of this character
vector defines the number of states.}

\item{sources}{Indices of source states.}

\item{terminal}{Indices of terminal states}

\item{self_loops}{Add self-loops to non-terminal and non-source states?}
}
\value{
A \code{\link{TransitionMatrix}}
}
\description{
\itemize{
\item \code{full}: All possible transitions
between non-terminal and non-source states are added, as well as transition
to terminal states from all states.
\item \code{survival}: basic survival model with single transition
\item \code{comprisk}: competing risks transition matrix
\item \code{illnessdeath}: Illness-Death model
\item \code{diamond}: Two intermediate states both leading to same terminal
state
}
}
