% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{transmat_comprisk}
\alias{transmat_comprisk}
\title{Create a competing risks transition matrix}
\usage{
transmat_comprisk(state_names = LETTERS[1:4])
}
\arguments{
\item{state_names}{Names of the states. The length of this character
vector defines the number of states.}
}
\value{
A \code{\link{TransitionMatrix}}.
}
\description{
Create a competing risks transition matrix
}
