% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{transmat_full}
\alias{transmat_full}
\title{Create "full" transition matrix}
\usage{
transmat_full(
  state_names = LETTERS[1:4],
  sources = 1,
  terminal = length(state_names),
  self_loops = TRUE
)
}
\arguments{
\item{state_names}{Names of the states. The length of this character
vector defines the number of states.}

\item{sources}{Indices of source states.}

\item{terminal}{Indices of terminal states}

\item{self_loops}{Add self-loops to non-terminal and non-source states?}
}
\value{
A \code{\link{TransitionMatrix}}, where all possible transitions
between non-terminal and non-source states are added, as well as transition
to terminal states from all states.
}
\description{
Create "full" transition matrix
}
