% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{transmat_illnessdeath}
\alias{transmat_illnessdeath}
\title{Illness-death model transition matrix}
\usage{
transmat_illnessdeath(state_names = c("Randomization", "Illness", "Death"))
}
\arguments{
\item{state_names}{Names of the states.}
}
\value{
A \code{\link{TransitionMatrix}}
}
\description{
Illness-death model transition matrix
}
