% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{transmat_survival}
\alias{transmat_survival}
\title{Basic alive-dead survival model transition matrix}
\usage{
transmat_survival(state_names = c("Alive", "Dead"))
}
\value{
A \code{\link{TransitionMatrix}}
}
\description{
Basic alive-dead survival model transition matrix
}
