% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{as_single_event}
\alias{as_single_event}
\title{PathData to time-to-event data format with a single event}
\usage{
as_single_event(pd, event)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object}

\item{event}{Name of the state corresponding to the event of interest (character)}
}
\value{
A \code{\link{PathData}} object
}
\description{
PathData to time-to-event data format with a single event
}
