% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{as_survival}
\alias{as_survival}
\title{PathData to event-free survival format}
\usage{
as_survival(pd, event)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object}

\item{event}{Name of the state corresponding to the event of interest (character)}
}
\value{
A \code{\link{PathData}} object
}
\description{
PathData to event-free survival format
}
\seealso{
Other PathData mutation functions: 
\code{\link{as_any_event}()},
\code{\link{as_single_event}()}
}
\concept{PathData mutation functions}
