% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{bmstate-package}
\alias{bmstate-package}
\alias{bmstate}
\title{The 'bmstate' package.}
\description{
Bayesian multistate modeling of right-censored time-to-event data.
}
\section{Transition matrix}{

For any analysis with this package, you will need a \code{\link{TransitionMatrix}}
which specifies the states and possible transitions between them.
This can be created directly with
\code{TransitionMatrix$new()} or using built-in functions for common
multistate systems such as \code{\link{transmat_illnessdeath}}.
}

\section{Data format}{

See the function \code{\link{df_to_pathdata}} for how to create a
\code{\link{PathData}} object from your data frame and a given
\code{\link{TransitionMatrix}}.
}

\section{Creating a model}{

You can create a model using
\code{\link{create_msm}}.
A model is represented by a \code{\link{MultistateModel}}
object, which has a \code{\link{MultistateSystem}}, which has a
\code{\link{TransitionMatrix}}.
}

\section{Fitting a model}{

You can fit a model with 'Stan' using \code{\link{fit_stan}}.
}

\section{Prediction with a fitted model}{


\itemize{
  \item For predicting state visit risks over a time period using state path
 simulation, see the functions \code{\link{generate_paths}},
 \code{\link{p_state_visit}}, and \code{\link{p_state_visit_per_subject}}.
  \item For analytically solving state occupancy probabilities after some time,
  see \code{\link{p_state_occupancy}}.
}
}

\section{Vignettes}{

See the vignettes for more information.
}

\author{
Juho Timonen (first.last at iki.fi) and Jacqueline Buros
}
\keyword{Bayesian}
\keyword{Stan}
\keyword{multistate}
