% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{bmstate-package}
\alias{bmstate-package}
\alias{bmstate}
\title{The 'bmstate' package.}
\description{
Bayesian multistate modeling using 'Stan'.
}
\section{Getting started}{

See the vignettes. You can create a model using \code{\link{create_msm}}
and fit one using \code{\link{fit_stan}}.
The following classes and their documentation can be useful

\itemize{
 \item \code{\link{MultistateModel}}
 \item \code{\link{PKModel}}
 \item \code{\link{PathData}}
 \item \code{\link{TransitionMatrix}}.
}
}

\author{
Juho Timonen (first.last at iki.fi) and Jacqueline Buros
}
\keyword{Bayesian}
\keyword{Stan}
\keyword{multistate}
