% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{bmstate-package}
\alias{bmstate-package}
\alias{bmstate}
\title{The 'bmstate' package.}
\description{
Bayesian multistate modeling using 'Stan'.
}
\section{Getting started}{

See the vignettes and the following exported functions and classes

\itemize{
 \item \code{\link{PathData}}
 \item \code{\link{create_brier_score_plot}}
 \item \code{\link{create_cindex_plot}}
 \item \code{\link{create_stan_data}}
 \item \code{\link{create_stan_model}}
 \item \code{\link{do_split}}
 \item \code{\link{fit_coxph}}
 \item \code{\link{fit_mstate}}
 \item \code{\link{generate_paths_many_subjects}}
 \item \code{\link{legend_to_PT_matrix}}
 \item \code{\link{legend_to_TFI_matrix}}
 \item \code{\link{plot_cumhaz_msfit}}
 \item \code{\link{plot_other_beta}}
 \item \code{\link{plot_p_event}}
 \item \code{\link{plot_p_event_by_dose}}
 \item \code{\link{simulate_example_data}}
 \item \code{\link{subject_df_with_idx}}
 \item \code{\link{summarize_ppsurv}}
}
}

\author{
Juho Timonen (first.last at iki.fi) and Jacki Buros
}
\keyword{Bayesian}
\keyword{Stan}
\keyword{multistate}
