% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{c_index}
\alias{c_index}
\title{Compute C index}
\usage{
c_index(pd, risk)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object of observed paths}

\item{risk}{Predicted event risk for each subject}
}
\description{
Compute C index
}
