% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{c_index}
\alias{c_index}
\title{Compute concordance index}
\usage{
c_index(df)
}
\arguments{
\item{df}{A \code{data.frame} with \code{surv} and \code{pred_prob} columns}
}
\description{
Compute concordance index
}
