% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModel.R
\name{create_msm}
\alias{create_msm}
\title{Create a multistate model}
\usage{
create_msm(matrix, states, hazard_covs, pk_covs = NULL, NK = 3, compile = TRUE)
}
\arguments{
\item{matrix}{A binary matrix where \code{matrix[i,j]} is 1 if transition
from state \code{i} to \code{j} is possible.}

\item{states}{A character vector of state names.}

\item{hazard_covs}{Covariates that affect the hazard. A character vector.}

\item{pk_covs}{Covariates that affect the PK parameters. A list with
elements \code{ka} \code{CL}, and \code{V2}. If \code{NULL}, a PK model
will not be created.}

\item{NK}{number of internal spline knots}

\item{compile}{compile 'Stan' model?}
}
\value{
A \code{\link{MultistateModel}} object.
}
\description{
Create a multistate model
}
