% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModel.R
\name{create_msm}
\alias{create_msm}
\title{Create a multistate model}
\usage{
create_msm(tm, hazard_covs, pk_covs = NULL, compile = TRUE)
}
\arguments{
\item{tm}{A \code{\link{TransitionMatrix}}}

\item{hazard_covs}{Covariates that affect the hazard. A character vector.}

\item{pk_covs}{Covariates that affect the PK parameters. A list with
elements \code{ka} \code{CL}, and \code{V2}. If \code{NULL}, a PK model
will not be created.}

\item{compile}{compile 'Stan' model?}
}
\value{
A \code{\link{MultistateModel}} object.
}
\description{
Create a multistate model
}
