% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Creating Stan data list}
\usage{
create_stan_data(
  model,
  data,
  dosing = NULL,
  prior_only = FALSE,
  delta_grid = 1
)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}} object.}

\item{data}{A \code{\link{JointData}} object of observed paths and dosing.}

\item{prior_only}{Sample from prior only?}

\item{delta_grid}{Time discretization delta for numerically integrating
hazards.}
}
\value{
A list of data for Stan.
}
\description{
Creating Stan data list
}
