% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Creating Stan data list}
\usage{
create_stan_data(model, data)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}} object.}

\item{data}{A \code{\link{JointData}} or \code{\link{PathData}} object.}
}
\value{
A list of data for Stan.
}
\description{
Creating Stan data list
}
