% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standata.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Creating Stan data list}
\usage{
create_stan_data(model, pd, dosing = NULL, split = NULL, prior_only = FALSE)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}} object.}

\item{pd}{A \code{\link{PathData}} object of observed paths.}

\item{dosing}{Dosing data frame.}

\item{split}{Train-test split. See \code{\link{do_split}}. If \code{NULL},
all subjects are training data.}

\item{prior_only}{Sample from prior only.}
}
\value{
A list of data for Stan.
}
\description{
Creating Stan data list
}
