% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Creating Stan data list}
\usage{
create_stan_data(model, data, prior_only = FALSE, pk_only = FALSE)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}} object.}

\item{data}{A \code{\link{JointData}} or \code{\link{PathData}} object.}

\item{prior_only}{Sample from prior only?}

\item{pk_only}{Do not fit hazard model parameters?}
}
\value{
A list of data for Stan.
}
\description{
Creating Stan data list
}
