% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standata.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Creating Stan data list}
\usage{
create_stan_data(
  pd,
  pk,
  covariates,
  ka_covariates,
  CL_covariates,
  V2_covariates,
  train_sub,
  test_sub,
  P = 30,
  NK = 3,
  t_max = NULL,
  do_pk = TRUE,
  prior_only = FALSE
)
}
\arguments{
\item{pd}{PD data frame}

\item{pk}{PK data frame}

\item{covariates}{hazard covariates}

\item{ka_covariates}{covariates that affect ka in PK model}

\item{CL_covariates}{covariates that affect CL in PK model}

\item{V2_covariates}{covariates that affect V2 in PK model}

\item{train_sub}{training subjects}

\item{test_sub}{test subjects}

\item{P}{num of prediction points}

\item{NK}{number of internal spline knots}

\item{t_max}{max time of interest}

\item{do_pk}{should pk model be used}

\item{prior_only}{ignore likelihood?}
}
\value{
A list of data for Stan.
}
\description{
Creating Stan data list
}
