% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{create_stan_model}
\alias{create_stan_model}
\title{Create the main 'Stan' model}
\usage{
create_stan_model(filepath = NULL, ...)
}
\arguments{
\item{filepath}{Deprecated.}

\item{...}{Arguments passed to 'CmdStanR'}
}
\description{
Uses the Stan code at
at \code{\link{default_stan_filepath}} or a custom file if it has been
set using \code{options(bmstate_stan_file = ...)}.
}
\seealso{
Other Stan-related functions: 
\code{\link{default_stan_filepath}()},
\code{\link{ensure_exposed_stan_functions}()},
\code{\link{fit_stan}()},
\code{\link{plot_stan_data_integral}()},
\code{\link{plot_stan_data_matrix}()}
}
\concept{Stan-related functions}
