% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{create_stan_model}
\alias{create_stan_model}
\title{Creates the Stan model}
\usage{
create_stan_model(mixp = FALSE)
}
\arguments{
\item{mixp}{use mixture prior?}
}
\value{
A CmdstanR model
}
\description{
Creates the Stan model
}
