% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{create_stan_model}
\alias{create_stan_model}
\title{Create the main 'Stan' model}
\usage{
create_stan_model(filepath = NULL, ...)
}
\arguments{
\item{filepath}{File path to the \code{msm.stan} file.}

\item{...}{Arguments passed to 'CmdStanR'}
}
\description{
Create the main 'Stan' model
}
