% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{df_to_pathdata}
\alias{df_to_pathdata}
\title{Data frame of one observed path per subject to PathData}
\usage{
df_to_pathdata(df, tm, covs)
}
\arguments{
\item{df}{Data frame}

\item{tm}{A transition matrix}

\item{covs}{covariates}
}
\value{
A \code{\link{PathData}} object
}
\description{
Data frame of one observed path per subject to PathData
}
