% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{df_to_pathdata}
\alias{df_to_pathdata}
\title{Create 'PathData' from a data frame of one observed path per subject}
\usage{
df_to_pathdata(df, tm, covs = NULL, validate = TRUE)
}
\arguments{
\item{df}{Data frame. Should have columns
\itemize{
  \item \code{subject_id} (character)
  \item \code{time} (numeric)
  \item \code{state} (integer, same indexing as in \code{tm})
  \item \code{is_transition} (logical)
  \item all the columns specified in \code{covs}
}
Rules:
\itemize{
   \item For each subject, there should be at least two rows, and the rows should
   be contiguous and the time should be non-decreasing.
   \item The \code{is_transition} value should indicate whether the row
   corresponds to a transition.
   \item The first row for each subject should never be a transition.
   \item For each row that is a transition, the transition from the state
   of the previous row to the current state should be a valid transition
   in \code{tm}.
   \item For each row that is not a transition, the state should not change
   from the previous row of that subject.
}}

\item{tm}{A \code{\link{TransitionMatrix}}}

\item{covs}{covariates (character vector)}

\item{validate}{Do stricter data validation? Recommended to use \code{TRUE}.}
}
\value{
A \code{\link{PathData}} object
}
\description{
Create 'PathData' from a data frame of one observed path per subject
}
