% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_split}
\alias{do_split}
\title{Train-test split}
\usage{
do_split(pd, p_test = 0.25)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object}

\item{p_test}{Proportion of test subjects}
}
\value{
a list with test and train subject indices
}
\description{
Train-test split
}
