% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{ensure_exposed_stan_functions}
\alias{ensure_exposed_stan_functions}
\title{Expose 'Stan' functions if they are not yet exposed}
\usage{
ensure_exposed_stan_functions()
}
\value{
logical telling whether they were already exposed
}
\description{
Expose 'Stan' functions if they are not yet exposed
}
