% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Fit a model using 'Stan'}
\usage{
fit_stan(model, data, prior_only = FALSE, delta_grid = 1, ...)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}} object.}

\item{data}{A \code{\link{JointData}} object of observed paths and dosing.}

\item{prior_only}{Sample from prior only?}

\item{delta_grid}{Time discretization delta for numerically integrating
hazards.}

\item{...}{Arguments passed to \code{sample} method of the
'CmdStanR' model.}
}
\value{
A \code{\link{MultistateModelFit}} object.
}
\description{
Fit a model using 'Stan'
}
