% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Fit a model using 'Stan'}
\usage{
fit_stan(
  model,
  data,
  prior_only = FALSE,
  pk_only = FALSE,
  set_auc_normalizers = TRUE,
  filepath = NULL,
  return_stanfit = FALSE,
  pathfinder = FALSE,
  ...
)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}} object.}

\item{data}{A \code{\link{JointData}} object of observed paths and dosing.}

\item{prior_only}{Sample from prior only?}

\item{pk_only}{Do not fit hazard model parameters?}

\item{set_auc_normalizers}{Set AUC normalization based on SS doses.}

\item{filepath}{Passed to \code{\link{create_stan_model}}.}

\item{return_stanfit}{Return also the raw 'Stan' fit object?}

\item{pathfinder}{Use pathfinder instead of MCMC?}

\item{...}{Arguments passed to \code{sample} method of the
'CmdStanR' model.}
}
\value{
A \code{\link{MultistateModelFit}} object.
}
\description{
**NOTE:** This function has a side effect of setting normalizers.
}
