% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Fit a model using 'Stan'}
\usage{
fit_stan(
  model,
  data,
  set_auc_normalizers = TRUE,
  filepath = NULL,
  return_stanfit = FALSE,
  method = "sample",
  ...
)
}
\arguments{
\item{model}{A \code{\link{MultistateModel}} object.}

\item{data}{A \code{\link{JointData}} or \code{\link{PathData}} object.}

\item{set_auc_normalizers}{Set AUC normalization based on SS doses.}

\item{filepath}{Deprecated.}

\item{return_stanfit}{Return also the raw 'Stan' fit object?}

\item{method}{Must be one of \code{"sample"} (default),
\code{"pathfinder"} or \code{"optimize"}.}

\item{...}{Arguments passed to the \code{sample},
\code{pathfinder} or \code{optimize}
method of the 'CmdStanR' model.}
}
\value{
A \code{\link{MultistateModelFit}} object.
}
\description{
\emph{NOTE:} This function has a side effect of setting covariate
normalizers and prior assumed mean baseline hazard based on data.
}
\seealso{
Other Stan-related functions: 
\code{\link{create_stan_data}()},
\code{\link{create_stan_model}()},
\code{\link{default_stan_filepath}()},
\code{\link{ensure_exposed_stan_functions}()},
\code{\link{plot_stan_data_integral}()},
\code{\link{plot_stan_data_matrix}()}
}
\concept{Stan-related functions}
