% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelStanFit.R
\name{generate_paths}
\alias{generate_paths}
\title{Path generation for 'MultistateModelStanFit'}
\usage{
generate_paths(fit, init_state = 1, t_max = NULL)
}
\arguments{
\item{init_state}{Index of starting state}

\item{t_max}{Max time (start time is always 0). If \code{NULL}, the max
time of the model is used.}
}
\value{
A \code{\link{PathData}} object.
}
\description{
Path generation for 'MultistateModelStanFit'
}
