% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelStanFit.R
\name{generate_paths}
\alias{generate_paths}
\title{Path generation for 'MultistateModelStanFit'}
\usage{
generate_paths(fit, init_state = 1, t_max = NULL, n_rep = 10, data = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelStanFit}} object}

\item{init_state}{Index of starting state}

\item{t_max}{Max time (start time is always 0). If \code{NULL}, the max
time of the model is used.}

\item{n_rep}{Number of repeats per draw.}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used. If not \code{NULL}, out-of-sample
mode is assumed (new subjects).}
}
\value{
A \code{\link{PathData}} object.
}
\description{
Path generation for 'MultistateModelStanFit'
}
