% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointData.R
\name{JointData}
\alias{JointData}
\title{Joint data class (R6 class)}
\description{
Contains data about both the state paths and dosing.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{paths}}{The events data, an object of class \code{\link{PathData}}.}

\item{\code{dosing}}{The dosing data, an object of class \code{\link{DosingData}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JointData-new}{\code{JointData$new()}}
\item \href{#method-JointData-filter}{\code{JointData$filter()}}
\item \href{#method-JointData-print}{\code{JointData$print()}}
\item \href{#method-JointData-plot_dosing}{\code{JointData$plot_dosing()}}
\item \href{#method-JointData-clone}{\code{JointData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointData-new"></a>}}
\if{latex}{\out{\hypertarget{method-JointData-new}{}}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointData$new(paths, dosing)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paths}}{The events data, an object of class \code{\link{PathData}}.}

\item{\code{dosing}}{The dosing data, an object of class \code{\link{DosingData}}.
Can be also \code{NULL}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointData-filter"></a>}}
\if{latex}{\out{\hypertarget{method-JointData-filter}{}}}
\subsection{Method \code{filter()}}{
Filter subjects, creates new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointData$filter(subject_ids_keep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject_ids_keep}}{Subjects to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointData-print"></a>}}
\if{latex}{\out{\hypertarget{method-JointData-print}{}}}
\subsection{Method \code{print()}}{
Print info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointData-plot_dosing"></a>}}
\if{latex}{\out{\hypertarget{method-JointData-plot_dosing}{}}}
\subsection{Method \code{plot_dosing()}}{
Plot dosing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointData$plot_dosing(df_fit = NULL, max_num_subjects = 12)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_fit}}{Fit data frame.}

\item{\code{max_num_subjects}}{Max number of subjects to plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{ggplot}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JointData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
