% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{msmfit_exposure}
\alias{msmfit_exposure}
\title{Compute exposure}
\usage{
msmfit_exposure(fit, oos = FALSE, data = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{oos}{Out-of-sample mode? If \code{FALSE}, the possible subject-specific
fitted parameters are used. If \code{TRUE}, acting
as if the subjects are new.}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used.}
}
\description{
Compute exposure
}
