% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{msmsf_log_hazard_multipliers}
\alias{msmsf_log_hazard_multipliers}
\title{Compute log_hazard multipliers}
\usage{
msmsf_log_hazard_multipliers(fit, data = NULL)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used. If not \code{NULL}, out-of-sample
mode is assumed (new subjects).}
}
\value{
A list of length \code{n_draws} where each element is a
matrix of shape \code{n_subject} x \code{n_transitions}
}
\description{
Compute log_hazard multipliers
}
