% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModel.R
\name{MultistateModel}
\alias{MultistateModel}
\title{Main model class}
\description{
Main model class

Main model class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{transmat}}{The transition matrix}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultistateModel-new}{\code{MultistateModel$new()}}
\item \href{#method-MultistateModel-get_stan_model}{\code{MultistateModel$get_stan_model()}}
\item \href{#method-MultistateModel-compile}{\code{MultistateModel$compile()}}
\item \href{#method-MultistateModel-print}{\code{MultistateModel$print()}}
\item \href{#method-MultistateModel-clone}{\code{MultistateModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$new(transmat, compile = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transmat}}{A \code{\link{TransitionMatrix}}.}

\item{\code{compile}}{Should the 'Stan' model code be created and compiled.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-get_stan_model"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-get_stan_model}{}}}
\subsection{Method \code{get_stan_model()}}{
Get the underlying 'Stan' model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$get_stan_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-compile"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-compile}{}}}
\subsection{Method \code{compile()}}{
Create and compile the 'Stan' model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$compile(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \code{cmdstan_model}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated model object (invisibly).
Print the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
