% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModel.R
\name{MultistateModel}
\alias{MultistateModel}
\title{Main model class}
\description{
Main model class

Main model class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{transmat}}{The transition matrix}

\item{\code{pk_model}}{A \code{\link{PKModel}} or NULL}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultistateModel-new}{\code{MultistateModel$new()}}
\item \href{#method-MultistateModel-set_knots}{\code{MultistateModel$set_knots()}}
\item \href{#method-MultistateModel-get_knots}{\code{MultistateModel$get_knots()}}
\item \href{#method-MultistateModel-get_tmax}{\code{MultistateModel$get_tmax()}}
\item \href{#method-MultistateModel-print}{\code{MultistateModel$print()}}
\item \href{#method-MultistateModel-covs}{\code{MultistateModel$covs()}}
\item \href{#method-MultistateModel-log_baseline_hazard}{\code{MultistateModel$log_baseline_hazard()}}
\item \href{#method-MultistateModel-simulate}{\code{MultistateModel$simulate()}}
\item \href{#method-MultistateModel-generate_transition}{\code{MultistateModel$generate_transition()}}
\item \href{#method-MultistateModel-generate_path}{\code{MultistateModel$generate_path()}}
\item \href{#method-MultistateModel-get_stan_model}{\code{MultistateModel$get_stan_model()}}
\item \href{#method-MultistateModel-compile}{\code{MultistateModel$compile()}}
\item \href{#method-MultistateModel-create_standata}{\code{MultistateModel$create_standata()}}
\item \href{#method-MultistateModel-fit}{\code{MultistateModel$fit()}}
\item \href{#method-MultistateModel-clone}{\code{MultistateModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$new(transmat, covariates, NK, pk_model = NULL, compile = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transmat}}{A \code{\link{TransitionMatrix}}.}

\item{\code{covariates}}{The names of the hazard covariates.}

\item{\code{NK}}{number of internal spline knots}

\item{\code{pk_model}}{A \code{\link{PKModel}} or NULL.}

\item{\code{compile}}{Should the 'Stan' model code be created and compiled.}

\item{\code{P}}{number of prediction points}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-set_knots"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-set_knots}{}}}
\subsection{Method \code{set_knots()}}{
Set knot locations based on event times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$set_knots(t_max, t_event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_max}}{Max time}

\item{\code{t_event}}{Occurred event times}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-get_knots"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-get_knots}{}}}
\subsection{Method \code{get_knots()}}{
Get knot locations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$get_knots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-get_tmax"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-get_tmax}{}}}
\subsection{Method \code{get_tmax()}}{
Get max time set for the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$get_tmax()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a number
Print the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-covs"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-covs}{}}}
\subsection{Method \code{covs()}}{
Get the hazard covariates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$covs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-log_baseline_hazard"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-log_baseline_hazard}{}}}
\subsection{Method \code{log_baseline_hazard()}}{
Evaluate log baseline hazard
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$log_baseline_hazard(t, log_w0, w = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{Output time points.}

\item{\code{log_w0}}{Intercept (log_scale).}

\item{\code{w}}{Spline weights (log scale). If \code{NULL}, will be set to
a vector of zeros, meaning that the log hazard is constant at
\code{log_w0}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector with same length as \code{t}
Generate paths
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-simulate}{}}}
\subsection{Method \code{simulate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$simulate()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w}}{An array of shape \code{n_paths} x \code{n_trans} x
\code{n_weights}}

\item{\code{log_w0}}{An array of shape \code{n_paths} x \code{n_trans}}

\item{\code{log_m}}{An array of shape \code{n_paths} x \code{n_trans}
Generate transition given state and transition intensity functions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-generate_transition"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-generate_transition}{}}}
\subsection{Method \code{generate_transition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$generate_transition(
  state,
  t_init,
  t_max,
  log_h0,
  m_sub,
  t_sub,
  UB = NULL,
  t_pred = NULL,
  tol = 1.03
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{Integer index of current state.}

\item{\code{t_init}}{Current time}

\item{\code{t_max}}{max time}

\item{\code{discretize}}{Discretize times to one time unit?
Generate a path starting from time 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-generate_path"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-generate_path}{}}}
\subsection{Method \code{generate_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$generate_path(init_state = 1, discretize = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{init_state}}{Integer index of starting state.}

\item{\code{discretize}}{Discretize times to one time unit?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-get_stan_model"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-get_stan_model}{}}}
\subsection{Method \code{get_stan_model()}}{
Get the underlying 'Stan' model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$get_stan_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-compile"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-compile}{}}}
\subsection{Method \code{compile()}}{
Create and compile the 'Stan' model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$compile(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \code{cmdstan_model}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated model object (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-create_standata"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-create_standata}{}}}
\subsection{Method \code{create_standata()}}{
Create the 'Stan' data list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$create_standata(pd, prior_only = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pd}}{A \code{\link{PathData}} object.}

\item{\code{prior_only}}{Sample from prior only?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$fit(pd, prior_only = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pd}}{A \code{\link{PathData}} object of observed paths.}

\item{\code{prior_only}}{Sample from prior only.}

\item{\code{...}}{Arguments passed to \code{sample} method of the
'CmdStanR' model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{MultistateModelFit}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
