% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModel.R
\name{MultistateModel}
\alias{MultistateModel}
\title{Main model class}
\description{
Main model class

Main model class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{system}}{A \code{\link{MultistateSystem}}}

\item{\code{pk_model}}{\emph{Experimental}. A \code{\link{PKModel}} or NULL.}

\item{\code{n_grid}}{Number of time discretization grid points for numerically
integrating hazards.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultistateModel-get_normalizers}{\code{MultistateModel$get_normalizers()}}
\item \href{#method-MultistateModel-get_auc_normalizers}{\code{MultistateModel$get_auc_normalizers()}}
\item \href{#method-MultistateModel-set_normalizers}{\code{MultistateModel$set_normalizers()}}
\item \href{#method-MultistateModel-set_auc_normalizers}{\code{MultistateModel$set_auc_normalizers()}}
\item \href{#method-MultistateModel-new}{\code{MultistateModel$new()}}
\item \href{#method-MultistateModel-set_knots}{\code{MultistateModel$set_knots()}}
\item \href{#method-MultistateModel-get_knots}{\code{MultistateModel$get_knots()}}
\item \href{#method-MultistateModel-get_tmax}{\code{MultistateModel$get_tmax()}}
\item \href{#method-MultistateModel-get_states}{\code{MultistateModel$get_states()}}
\item \href{#method-MultistateModel-has_pk}{\code{MultistateModel$has_pk()}}
\item \href{#method-MultistateModel-print}{\code{MultistateModel$print()}}
\item \href{#method-MultistateModel-covs}{\code{MultistateModel$covs()}}
\item \href{#method-MultistateModel-data_covs}{\code{MultistateModel$data_covs()}}
\item \href{#method-MultistateModel-categ_covs}{\code{MultistateModel$categ_covs()}}
\item \href{#method-MultistateModel-simulate_data}{\code{MultistateModel$simulate_data()}}
\item \href{#method-MultistateModel-simulate_pk_data}{\code{MultistateModel$simulate_pk_data()}}
\item \href{#method-MultistateModel-simulate_events}{\code{MultistateModel$simulate_events()}}
\item \href{#method-MultistateModel-simulate_subjects}{\code{MultistateModel$simulate_subjects()}}
\item \href{#method-MultistateModel-target_states}{\code{MultistateModel$target_states()}}
\item \href{#method-MultistateModel-clone}{\code{MultistateModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-get_normalizers"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-get_normalizers}{}}}
\subsection{Method \code{get_normalizers()}}{
Get normalization constants for each variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$get_normalizers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-get_auc_normalizers"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-get_auc_normalizers}{}}}
\subsection{Method \code{get_auc_normalizers()}}{
Get normalization constants for AUC (PK)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$get_auc_normalizers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
Set normalization constant for each variable (side effect)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-set_normalizers"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-set_normalizers}{}}}
\subsection{Method \code{set_normalizers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$set_normalizers(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \code{\link{JointData}} object
Set normalization constants for AUC (side effect)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-set_auc_normalizers"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-set_auc_normalizers}{}}}
\subsection{Method \code{set_auc_normalizers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$set_auc_normalizers(loc = 0, scale = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loc}}{Location}

\item{\code{scale}}{Scale}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$new(
  system,
  covariates = NULL,
  pk_model = NULL,
  tmax = 1000,
  num_knots = 5,
  categorical = NULL,
  n_grid = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{system}}{A \code{\link{MultistateSystem}}}

\item{\code{covariates}}{The names of the hazard covariates (excluding possible
exposure estimated from PK model). Do not use reserved names
\code{ss_auc} or \code{dose}.}

\item{\code{pk_model}}{A \code{\link{PKModel}} or NULL.}

\item{\code{tmax}}{Max time.}

\item{\code{num_knots}}{Total number of spline knots.}

\item{\code{categorical}}{Names of categorical covariates.}

\item{\code{n_grid}}{Number of time discretization points for integrating
hazards.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-set_knots"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-set_knots}{}}}
\subsection{Method \code{set_knots()}}{
Set knot locations based on event times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$set_knots(t_max, t_event, num_knots)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_max}}{Max time}

\item{\code{t_event}}{Occurred event times}

\item{\code{num_knots}}{Total number of knots}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-get_knots"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-get_knots}{}}}
\subsection{Method \code{get_knots()}}{
Get knots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$get_knots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-get_tmax"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-get_tmax}{}}}
\subsection{Method \code{get_tmax()}}{
Get max time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$get_tmax()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-get_states"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-get_states}{}}}
\subsection{Method \code{get_states()}}{
Get names of the states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$get_states()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-has_pk"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-has_pk}{}}}
\subsection{Method \code{has_pk()}}{
Is there a PK submodel?
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$has_pk()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-covs"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-covs}{}}}
\subsection{Method \code{covs()}}{
Get the hazard covariates (including steady-state exposure
if PK model is included)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$covs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-data_covs"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-data_covs}{}}}
\subsection{Method \code{data_covs()}}{
Get all covariates that need to be given as data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$data_covs(which = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which}}{Which subset to get?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-categ_covs"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-categ_covs}{}}}
\subsection{Method \code{categ_covs()}}{
Get names of categorical covariates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$categ_covs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-simulate_data"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-simulate_data}{}}}
\subsection{Method \code{simulate_data()}}{
Simulate data using the multistate model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$simulate_data(
  N_subject = 100,
  beta_haz = NULL,
  beta_pk = NULL,
  w0 = 0.001,
  w = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N_subject}}{Number of subjects.}

\item{\code{beta_haz}}{Covariate effects on each transition type.
A matrix of shape \code{num_target_states} x \code{num_covs}.
If \code{NULL}, a data frame of zeros is used.}

\item{\code{beta_pk}}{Covariate effects on PK parameters. A named list with
three elements, each being a vector. If any element is \code{NULL},
a vector of zeros is used.}

\item{\code{w0}}{Baseline hazard rate for all transitions}

\item{\code{w}}{Spline weights. Matrix of shape \code{num_trans} x
\code{num_weights}. If \code{NULL}, a matrix of zeros is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{JointData}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-simulate_pk_data"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-simulate_pk_data}{}}}
\subsection{Method \code{simulate_pk_data()}}{
Simulate PK data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$simulate_pk_data(df_subjects, beta_pk = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_subjects}}{The subjects data frame}

\item{\code{beta_pk}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{data.frame} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-simulate_events"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-simulate_events}{}}}
\subsection{Method \code{simulate_events()}}{
Simulate events data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$simulate_events(df_subjects, beta_haz, log_w0, w)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_subjects}}{The subjects data frame}

\item{\code{beta_haz}}{Matrix of shape \code{num_target_states} x \code{num_covs}}

\item{\code{log_w0}}{Baseline log hazard rate, vector with length
\code{num_trans}}

\item{\code{w}}{Spline weights. Matrix of shape \code{num_trans} x
\code{num_weights}. If \code{NULL}, a matrix of zeros is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{tibble}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-simulate_subjects"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-simulate_subjects}{}}}
\subsection{Method \code{simulate_subjects()}}{
Simulate subject data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$simulate_subjects(N_subject = 100, doses = c(15, 30, 60))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N_subject}}{Number of subjects.}

\item{\code{doses}}{Possible doses. Only has effect if a PK submodel exists.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble}
Get indices of states that are not source states
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-target_states"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-target_states}{}}}
\subsection{Method \code{target_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$target_states()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
