% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{MultistateModelFit}
\alias{MultistateModelFit}
\title{The Fit class}
\description{
The Fit class

The Fit class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultistateModelFit-get_data}{\code{MultistateModelFit$get_data()}}
\item \href{#method-MultistateModelFit-get_model}{\code{MultistateModelFit$get_model()}}
\item \href{#method-MultistateModelFit-new}{\code{MultistateModelFit$new()}}
\item \href{#method-MultistateModelFit-get_stan_fit}{\code{MultistateModelFit$get_stan_fit()}}
\item \href{#method-MultistateModelFit-get_stan_data}{\code{MultistateModelFit$get_stan_data()}}
\item \href{#method-MultistateModelFit-draws}{\code{MultistateModelFit$draws()}}
\item \href{#method-MultistateModelFit-log_z_pars}{\code{MultistateModelFit$log_z_pars()}}
\item \href{#method-MultistateModelFit-loglik}{\code{MultistateModelFit$loglik()}}
\item \href{#method-MultistateModelFit-gq}{\code{MultistateModelFit$gq()}}
\item \href{#method-MultistateModelFit-clone}{\code{MultistateModelFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get original data used to fit the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$get_data(dataname = "LON")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{Name of dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-get_model}{}}}
\subsection{Method \code{get_model()}}{
Get model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$get_model(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of model. Only has effect for
\code{\link{JointModelFit}} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-new}{}}}
\subsection{Method \code{new()}}{
Create model fit object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$new(model, stan_fit, datasets, stan_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A \code{\link{StanModel}} object.}

\item{\code{stan_fit}}{The created 'Stan' fit.}

\item{\code{datasets}}{Original data sets.}

\item{\code{stan_data}}{The created 'Stan' data. Stored mainly for easier
debugging.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-get_stan_fit"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-get_stan_fit}{}}}
\subsection{Method \code{get_stan_fit()}}{
Get the underlying 'Stan' fit object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$get_stan_fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-get_stan_data"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-get_stan_data}{}}}
\subsection{Method \code{get_stan_data()}}{
Get the underlying 'Stan' data object (for debugging).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$get_stan_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-draws"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-draws}{}}}
\subsection{Method \code{draws()}}{
Extract draws as \code{rvar}s
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$draws(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Param/quantity name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-log_z_pars"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-log_z_pars}{}}}
\subsection{Method \code{log_z_pars()}}{
Full names of parameters that start with 'log_z_'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$log_z_pars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-loglik"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-loglik}{}}}
\subsection{Method \code{loglik()}}{
Extract log likelihood as 'rvars'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$loglik()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-gq"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-gq}{}}}
\subsection{Method \code{gq()}}{
Generate quantities using the fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$gq(stan_data = NULL, fitted_params = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stan_data}}{Full 'Stan' input list.}

\item{\code{fitted_params}}{Argument to \code{generate_quantities()}.}

\item{\code{...}}{Other arguments to \code{generate_quantities()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
