% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{MultistateModelFit}
\alias{MultistateModelFit}
\title{Minimal fit class}
\description{
Minimal fit class

Minimal fit class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The \code{\link{MultistateModel}}}

\item{\code{data}}{A  \code{\link{JointData}} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultistateModelFit-new}{\code{MultistateModelFit$new()}}
\item \href{#method-MultistateModelFit-mean_fit}{\code{MultistateModelFit$mean_fit()}}
\item \href{#method-MultistateModelFit-is_point_estimate}{\code{MultistateModelFit$is_point_estimate()}}
\item \href{#method-MultistateModelFit-get_data}{\code{MultistateModelFit$get_data()}}
\item \href{#method-MultistateModelFit-get_draws}{\code{MultistateModelFit$get_draws()}}
\item \href{#method-MultistateModelFit-get_draws_of}{\code{MultistateModelFit$get_draws_of()}}
\item \href{#method-MultistateModelFit-print}{\code{MultistateModelFit$print()}}
\item \href{#method-MultistateModelFit-plot_basisfun}{\code{MultistateModelFit$plot_basisfun()}}
\item \href{#method-MultistateModelFit-plot_pk}{\code{MultistateModelFit$plot_pk()}}
\item \href{#method-MultistateModelFit-plot_h0}{\code{MultistateModelFit$plot_h0()}}
\item \href{#method-MultistateModelFit-h0_dist}{\code{MultistateModelFit$h0_dist()}}
\item \href{#method-MultistateModelFit-log_z_pars}{\code{MultistateModelFit$log_z_pars()}}
\item \href{#method-MultistateModelFit-num_draws}{\code{MultistateModelFit$num_draws()}}
\item \href{#method-MultistateModelFit-clone}{\code{MultistateModelFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-new}{}}}
\subsection{Method \code{new()}}{
Create model fit object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$new(data, stan_data, model, draws)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data used to create the fit.}

\item{\code{stan_data}}{The used 'Stan' data list.}

\item{\code{model}}{A \code{\link{MultistateModel}}}

\item{\code{draws}}{A named list of rvars.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-mean_fit"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-mean_fit}{}}}
\subsection{Method \code{mean_fit()}}{
Create a reduced version with only a single draw, corresponding
to the mean of original draws.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$mean_fit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{\link{MultistateModelFit}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-is_point_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-is_point_estimate}{}}}
\subsection{Method \code{is_point_estimate()}}{
Check if fit is a point estimate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$is_point_estimate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a logical value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-get_data}{}}}
\subsection{Method \code{get_data()}}{
Extract data list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-get_draws"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-get_draws}{}}}
\subsection{Method \code{get_draws()}}{
Extract draws as \code{rvar}s
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$get_draws(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Param/quantity name
Draws in a raw array with same shape as Stan variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-get_draws_of"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-get_draws_of}{}}}
\subsection{Method \code{get_draws_of()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$get_draws_of(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Param/quantity name of \code{x}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Array with dimension \code{c(ndraws(x), dim(x))}
Print the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-plot_basisfun"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-plot_basisfun}{}}}
\subsection{Method \code{plot_basisfun()}}{
Plot used basis functions (grid)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$plot_basisfun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-plot_pk"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-plot_pk}{}}}
\subsection{Method \code{plot_pk()}}{
Plot PK fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$plot_pk(
  max_num_subjects = 12,
  data = NULL,
  L = 30,
  timescale = 24,
  n_prev = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_num_subjects}}{Max number of subjects to show.}

\item{\code{data}}{Data for which to predict the concentration. If \code{NULL},
training data is used.}

\item{\code{L}}{number of grid points for each subject}

\item{\code{timescale}}{scale of time}

\item{\code{n_prev}}{number of previous doses to show fit for
Plot baseline hazard distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-plot_h0"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-plot_h0}{}}}
\subsection{Method \code{plot_h0()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$plot_h0(t = NULL, ci_alpha = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{times where to evaluate the baseline hazards}

\item{\code{ci_alpha}}{width of credible interval
Baseline hazard distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-h0_dist"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-h0_dist}{}}}
\subsection{Method \code{h0_dist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$h0_dist(t = NULL, ci_alpha = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{times where to evaluate the baseline hazards}

\item{\code{ci_alpha}}{width of credible interval}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-log_z_pars"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-log_z_pars}{}}}
\subsection{Method \code{log_z_pars()}}{
Full names of parameters that start with \code{log_z_}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$log_z_pars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-num_draws"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-num_draws}{}}}
\subsection{Method \code{num_draws()}}{
Get number of draws
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$num_draws()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
