% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelStanFit.R
\name{MultistateModelStanFit}
\alias{MultistateModelStanFit}
\title{Minimal fit class}
\description{
Minimal fit class

Minimal fit class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The \code{\link{MultistateModel}}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultistateModelStanFit-new}{\code{MultistateModelStanFit$new()}}
\item \href{#method-MultistateModelStanFit-get_data}{\code{MultistateModelStanFit$get_data()}}
\item \href{#method-MultistateModelStanFit-get_draws}{\code{MultistateModelStanFit$get_draws()}}
\item \href{#method-MultistateModelStanFit-get_draws_of}{\code{MultistateModelStanFit$get_draws_of()}}
\item \href{#method-MultistateModelStanFit-print}{\code{MultistateModelStanFit$print()}}
\item \href{#method-MultistateModelStanFit-plot_basisfun}{\code{MultistateModelStanFit$plot_basisfun()}}
\item \href{#method-MultistateModelStanFit-plot_h0}{\code{MultistateModelStanFit$plot_h0()}}
\item \href{#method-MultistateModelStanFit-h0_dist}{\code{MultistateModelStanFit$h0_dist()}}
\item \href{#method-MultistateModelStanFit-log_z_pars}{\code{MultistateModelStanFit$log_z_pars()}}
\item \href{#method-MultistateModelStanFit-num_draws}{\code{MultistateModelStanFit$num_draws()}}
\item \href{#method-MultistateModelStanFit-gq}{\code{MultistateModelStanFit$gq()}}
\item \href{#method-MultistateModelStanFit-clone}{\code{MultistateModelStanFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-new}{}}}
\subsection{Method \code{new()}}{
Create model fit object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$new(stan_fit, stan_data, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stan_fit}}{A 'Stan' fit object}

\item{\code{stan_data}}{The used 'Stan' data list.}

\item{\code{model}}{A \code{\link{MultistateModel}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-get_data}{}}}
\subsection{Method \code{get_data()}}{
Extract data list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-get_draws"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-get_draws}{}}}
\subsection{Method \code{get_draws()}}{
Extract draws as \code{rvar}s
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$get_draws(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Param/quantity name
Draws in a raw array with same shape as Stan variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-get_draws_of"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-get_draws_of}{}}}
\subsection{Method \code{get_draws_of()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$get_draws_of(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Param/quantity name of \code{x}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Array with dimension \code{c(ndraws(x), dim(x))}
Print the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-plot_basisfun"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-plot_basisfun}{}}}
\subsection{Method \code{plot_basisfun()}}{
Plot used basis functions (grid)
Plot baseline hazard distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$plot_basisfun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-plot_h0"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-plot_h0}{}}}
\subsection{Method \code{plot_h0()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$plot_h0(t = NULL, ci_alpha = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{times where to evaluate the baseline hazards}

\item{\code{ci_alpha}}{width of credible interval
Baseline hazard distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-h0_dist"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-h0_dist}{}}}
\subsection{Method \code{h0_dist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$h0_dist(t = NULL, ci_alpha = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{times where to evaluate the baseline hazards}

\item{\code{ci_alpha}}{width of credible interval}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-log_z_pars"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-log_z_pars}{}}}
\subsection{Method \code{log_z_pars()}}{
Full names of parameters that start with \code{log_z_}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$log_z_pars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-num_draws"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-num_draws}{}}}
\subsection{Method \code{num_draws()}}{
Get number of draws
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$num_draws()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-gq"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-gq}{}}}
\subsection{Method \code{gq()}}{
Generate quantities using the fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$gq(stan_data = NULL, fitted_params = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stan_data}}{Full 'Stan' input list.}

\item{\code{fitted_params}}{Argument to \code{generate_quantities()}.}

\item{\code{...}}{Other arguments to \code{generate_quantities()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateModelStanFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateModelStanFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateModelStanFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
