% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateSystem.R
\name{MultistateSystem}
\alias{MultistateSystem}
\title{A multistate system with proportional spline hazards}
\description{
A multistate system with proportional spline hazards

A multistate system with proportional spline hazards
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultistateSystem-new}{\code{MultistateSystem$new()}}
\item \href{#method-MultistateSystem-num_states}{\code{MultistateSystem$num_states()}}
\item \href{#method-MultistateSystem-num_trans}{\code{MultistateSystem$num_trans()}}
\item \href{#method-MultistateSystem-num_weights}{\code{MultistateSystem$num_weights()}}
\item \href{#method-MultistateSystem-num_knots}{\code{MultistateSystem$num_knots()}}
\item \href{#method-MultistateSystem-tm}{\code{MultistateSystem$tm()}}
\item \href{#method-MultistateSystem-knots_not_set}{\code{MultistateSystem$knots_not_set()}}
\item \href{#method-MultistateSystem-set_knots}{\code{MultistateSystem$set_knots()}}
\item \href{#method-MultistateSystem-get_knots}{\code{MultistateSystem$get_knots()}}
\item \href{#method-MultistateSystem-get_tmax}{\code{MultistateSystem$get_tmax()}}
\item \href{#method-MultistateSystem-print}{\code{MultistateSystem$print()}}
\item \href{#method-MultistateSystem-log_baseline_hazard}{\code{MultistateSystem$log_baseline_hazard()}}
\item \href{#method-MultistateSystem-basisfun_matrix}{\code{MultistateSystem$basisfun_matrix()}}
\item \href{#method-MultistateSystem-log_inst_hazard}{\code{MultistateSystem$log_inst_hazard()}}
\item \href{#method-MultistateSystem-max_inst_hazard}{\code{MultistateSystem$max_inst_hazard()}}
\item \href{#method-MultistateSystem-simulate}{\code{MultistateSystem$simulate()}}
\item \href{#method-MultistateSystem-clone}{\code{MultistateSystem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$new(transmat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transmat}}{A \code{\link{TransitionMatrix}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-num_states"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-num_states}{}}}
\subsection{Method \code{num_states()}}{
Get number of states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$num_states()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-num_trans"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-num_trans}{}}}
\subsection{Method \code{num_trans()}}{
Get number of transitions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$num_trans()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer
Get number of spline weight parameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-num_weights"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-num_weights}{}}}
\subsection{Method \code{num_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$num_weights()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an integer (\code{num_knots} - 2) + \code{self$spline_k}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-num_knots"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-num_knots}{}}}
\subsection{Method \code{num_knots()}}{
Get number of spline knots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$num_knots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer, zero if knots have not been set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-tm"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-tm}{}}}
\subsection{Method \code{tm()}}{
Get transition matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$tm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-knots_not_set"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-knots_not_set}{}}}
\subsection{Method \code{knots_not_set()}}{
Have knots not been set yet?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$knots_not_set()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a logical
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-set_knots"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-set_knots}{}}}
\subsection{Method \code{set_knots()}}{
Set spline knot locations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$set_knots(locations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{locations}}{A numeric vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-get_knots"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-get_knots}{}}}
\subsection{Method \code{get_knots()}}{
Get knot locations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$get_knots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-get_tmax"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-get_tmax}{}}}
\subsection{Method \code{get_tmax()}}{
Get max time set for the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$get_tmax()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a number
Print the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-log_baseline_hazard"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-log_baseline_hazard}{}}}
\subsection{Method \code{log_baseline_hazard()}}{
Evaluate log baseline hazard
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$log_baseline_hazard(t, log_w0, w = NULL, SBF = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{Output time points. Not used if \code{SBF} is given.}

\item{\code{log_w0}}{Intercept (log_scale).}

\item{\code{w}}{Spline weights (log scale). If \code{NULL}, will be set to
a vector of zeros, meaning that the log hazard is constant at
\code{log_w0}}

\item{\code{SBF}}{Precomputed basisfunction matrix at \code{t}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector with same length as \code{t}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-basisfun_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-basisfun_matrix}{}}}
\subsection{Method \code{basisfun_matrix()}}{
Evaluate basis function matrix for baseline hazard
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$basisfun_matrix(t)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{Evaluation time points.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix with shape \code{c(N, L+1)} where \code{L} is number of
knots
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-log_inst_hazard"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-log_inst_hazard}{}}}
\subsection{Method \code{log_inst_hazard()}}{
Evaluate log instant hazard
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$log_inst_hazard(t, w, log_w0, log_m, SBF = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{Time point(s). Not used if \code{SBF} is given.}

\item{\code{w}}{Spline basis function weights (vector)}

\item{\code{log_w0}}{Intercept (log)}

\item{\code{log_m}}{Hazard multiplier (log)}

\item{\code{SBF}}{Precomputed basisfunction matrix at \code{t}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-max_inst_hazard"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-max_inst_hazard}{}}}
\subsection{Method \code{max_inst_hazard()}}{
Max instant hazard on interval (t1, t2)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$max_inst_hazard(t1, t2, w, log_w0, log_m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t1}}{Start time point}

\item{\code{t2}}{End time point}

\item{\code{w}}{Spline basis function weights (vector)}

\item{\code{log_w0}}{Intercept (log)}

\item{\code{log_m}}{Hazard multiplier (log)
Generate paths}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-simulate}{}}}
\subsection{Method \code{simulate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$simulate(
  w,
  log_w0,
  log_m,
  init_state = 1,
  t_max = NULL,
  n_rep = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w}}{An array of shape \code{n_draws} x \code{n_trans} x
\code{n_weights}}

\item{\code{log_w0}}{An array of shape \code{n_draws} x \code{n_trans}}

\item{\code{log_m}}{An array of shape \code{n_draws} x \code{n_trans}}

\item{\code{init_state}}{Integer index of starting state.}

\item{\code{t_max}}{Max time. If not given, \code{self$get_tmax()} is used.}

\item{\code{n_rep}}{Number of repetitions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with \code{n_draws} x \code{n_rep} paths.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultistateSystem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultistateSystem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultistateSystem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
