% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultistateModelFit.R
\name{p_state_occupancy}
\alias{p_state_occupancy}
\title{Solve state occupancy probabilities for each subject and draw in
'MultistateModelFit'}
\usage{
p_state_occupancy(
  fit,
  oos = FALSE,
  t_start = 0,
  t_out = NULL,
  data = NULL,
  ...
)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{oos}{Out-of-sample mode? If \code{FALSE}, the possible subject-specific
fitted parameters are used. If \code{TRUE}, acting
as if the subjects are new.}

\item{t_start}{Start time. The results are transition probabilities to each
state, evaluated at each time point of \code{t_out}, given that the
that the subject is in the state they are in \code{data} at \code{t_start}.}

\item{t_out}{End times (vector)}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used.}

\item{...}{Arguments passed to \code{deSolve::ode()}.}
}
\value{
A data frame with columns
\itemize{
  \item \code{subject_id} (character)
   \item \code{time} (numeric)
   \item \code{state} (character)
   \item \code{prob} state occupancy probability for given subject at given time
 }
 Note that \code{prob} is an \code{rvar} over all parameter draws.
}
\description{
Solve state occupancy probabilities for each subject and draw in
'MultistateModelFit'
}
