% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{PathData}
\alias{PathData}
\title{Path data class (R6 class)}
\description{
Path data class (R6 class)

Path data class (R6 class)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subject_df}}{Data frame with one row per subject. Must have
\code{subject_id} and all covariates as columns.}

\item{\code{path_df}}{Data frame of actual paths. Must have \code{path_id},
\code{state}, \code{time}, and \code{is_event} as columns.}

\item{\code{link_df}}{Links the path and subject data frames. Must have
\code{path_id}, \code{draw_idx}, \code{rep_idx}, and \code{subject_id} as
columns.}

\item{\code{covs}}{Covariate column names.}

\item{\code{state_names}}{Names of the states (character vector).}

\item{\code{state_types}}{State types (integer vector).}

\item{\code{terminal_states}}{Names of terminal states.}

\item{\code{null_state}}{Names of null states. These are not events.}

\item{\code{censor_state}}{Names of censoring states.
Initialize}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PathData-new}{\code{PathData$new()}}
\item \href{#method-PathData-covariate_names}{\code{PathData$covariate_names()}}
\item \href{#method-PathData-lengths}{\code{PathData$lengths()}}
\item \href{#method-PathData-n_paths}{\code{PathData$n_paths()}}
\item \href{#method-PathData-longest_path}{\code{PathData$longest_path()}}
\item \href{#method-PathData-get_event_states}{\code{PathData$get_event_states()}}
\item \href{#method-PathData-get_event_state_names}{\code{PathData$get_event_state_names()}}
\item \href{#method-PathData-as_msdata}{\code{PathData$as_msdata()}}
\item \href{#method-PathData-as_time_to_first_event}{\code{PathData$as_time_to_first_event()}}
\item \href{#method-PathData-event_states}{\code{PathData$event_states()}}
\item \href{#method-PathData-print}{\code{PathData$print()}}
\item \href{#method-PathData-check_valid_state}{\code{PathData$check_valid_state()}}
\item \href{#method-PathData-check_states}{\code{PathData$check_states()}}
\item \href{#method-PathData-as_data_frame}{\code{PathData$as_data_frame()}}
\item \href{#method-PathData-as_transitions}{\code{PathData$as_transitions()}}
\item \href{#method-PathData-plot_paths}{\code{PathData$plot_paths()}}
\item \href{#method-PathData-trans_matrix}{\code{PathData$trans_matrix()}}
\item \href{#method-PathData-plot_graph}{\code{PathData$plot_graph()}}
\item \href{#method-PathData-coxph}{\code{PathData$coxph()}}
\item \href{#method-PathData-filter}{\code{PathData$filter()}}
\item \href{#method-PathData-clone}{\code{PathData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-new"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$new(
  subject_df,
  path_df,
  link_df,
  state_names,
  covs = NULL,
  check_order = TRUE,
  terminal_states = NULL,
  censor_state = "Censor",
  null_state = "Randomization"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject_df}}{Subjects data frame.}

\item{\code{path_df}}{Paths data frame.}

\item{\code{link_df}}{Link data frame.}

\item{\code{state_names}}{Names of all states.}

\item{\code{covs}}{Covariate column names.}

\item{\code{check_order}}{Check order of paths?}

\item{\code{terminal_states}}{Names of terminal states.}

\item{\code{censor_state}}{Name of censoring state.
Get names of covariates}

\item{\code{null_state}}{Name of null state.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-covariate_names"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-covariate_names}{}}}
\subsection{Method \code{covariate_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$covariate_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character vector
Get path lenghts (among paths that include events)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-lengths"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-lengths}{}}}
\subsection{Method \code{lengths()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$lengths()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a data frame of path ids and counts
Get number of paths
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-n_paths"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-n_paths}{}}}
\subsection{Method \code{n_paths()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$n_paths()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an integer
Get longest path
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-longest_path"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-longest_path}{}}}
\subsection{Method \code{longest_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$longest_path()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{\link{PathData}} object with just one path
Get indices of event states
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-get_event_states"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-get_event_states}{}}}
\subsection{Method \code{get_event_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$get_event_states()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer vector
Get names of states that are events (i.e. not initial and not censor)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-get_event_state_names"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-get_event_state_names}{}}}
\subsection{Method \code{get_event_state_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$get_event_state_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector
Convert to format used by the 'mstate' package
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_msdata"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_msdata}{}}}
\subsection{Method \code{as_msdata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_msdata(covariates = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Include covariates?
Format as time to first event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_time_to_first_event"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_time_to_first_event}{}}}
\subsection{Method \code{as_time_to_first_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_time_to_first_event(covs = c("subject_id"), truncate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covs}}{covariates to include in output data frame}

\item{\code{truncate}}{truncate after terminal events?
Get names of event states}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-event_states"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-event_states}{}}}
\subsection{Method \code{event_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$event_states()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character vector
Print info
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-print"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
Check that a state is valid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-check_valid_state"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-check_valid_state}{}}}
\subsection{Method \code{check_valid_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$check_valid_state(check_states, state_names, min_len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{check_states}}{states to check}

\item{\code{state_names}}{all state names}

\item{\code{min_len}}{minimum length
Check that states input is valid}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-check_states"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-check_states}{}}}
\subsection{Method \code{check_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$check_states(
  state_names,
  terminal_states,
  null_state,
  censor_state,
  path_df
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state_names}}{names of all states}

\item{\code{terminal_states}}{names of terminal states}

\item{\code{null_state}}{name of initial states}

\item{\code{censor_state}}{name of censoring states}

\item{\code{path_df}}{the path data frame
Convert to one long data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_data_frame}{}}}
\subsection{Method \code{as_data_frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_data_frame(covariates = NULL, truncate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Which covariates to include?}

\item{\code{truncate}}{Truncate after terminal events?
Data frame in transitions format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_transitions"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_transitions}{}}}
\subsection{Method \code{as_transitions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_transitions(only_observed = FALSE, force_rerun = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_observed}}{Limit transitions to only the observed ones?}

\item{\code{force_rerun}}{Force rerun of the conversion? If \code{FALSE},
a cached version is used if it exists.
Step plot of the paths}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-plot_paths"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-plot_paths}{}}}
\subsection{Method \code{plot_paths()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$plot_paths(n_paths = NULL, alpha = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_paths}}{Number of paths to subsample for plotting.}

\item{\code{alpha}}{opacity
Transition proportion matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-trans_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-trans_matrix}{}}}
\subsection{Method \code{trans_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$trans_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a matrix
Visualize the transition proportion matrix as a graph
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-plot_graph"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-plot_graph}{}}}
\subsection{Method \code{plot_graph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$plot_graph(digits = 3, include_censor = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{Max number of digits to show in numbers}

\item{\code{include_censor}}{Include censoring state in the graph?}

\item{\code{...}}{Arguments passed to \code{qgraph}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{qgraph} plot
Fit Cox proportional hazards model
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-coxph"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-coxph}{}}}
\subsection{Method \code{coxph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$coxph(covs = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covs}}{Covariate names.}

\item{\code{...}}{Arguments passed to \code{survival::coxph}.
Filter based on path id, creates new object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-filter"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-filter}{}}}
\subsection{Method \code{filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$filter(
  path_ids_keep = NULL,
  subject_ids_keep = NULL,
  rep_ids_keep = NULL,
  draw_ids_keep = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_ids_keep}}{Path ids to keep}

\item{\code{subject_ids_keep}}{Subject ids to keep}

\item{\code{rep_ids_keep}}{Repetition ids to keep}

\item{\code{draw_ids_keep}}{Draw ids to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
