% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{PathData}
\alias{PathData}
\title{Path data class (R6 class)}
\description{
Path data class (R6 class)

Path data class (R6 class)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subject_df}}{Data frame with one row per subject. Must have
\code{subject_id} and all covariates as columns.}

\item{\code{path_df}}{Data frame of actual paths. Must have \code{path_id},
\code{state}, \code{time}, and \code{is_event} as columns.}

\item{\code{link_df}}{Links the path and subject data frames. Must have
\code{path_id}, \code{draw_idx}, \code{rep_idx}, and \code{subject_id} as
columns.}

\item{\code{covs}}{Covariate column names.}

\item{\code{transmat}}{A \code{\link{TransitionMatrix}} describing the system in to
which the paths belong.
Initialize}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PathData-new}{\code{PathData$new()}}
\item \href{#method-PathData-covariate_names}{\code{PathData$covariate_names()}}
\item \href{#method-PathData-lengths}{\code{PathData$lengths()}}
\item \href{#method-PathData-n_paths}{\code{PathData$n_paths()}}
\item \href{#method-PathData-longest_path}{\code{PathData$longest_path()}}
\item \href{#method-PathData-censor_state}{\code{PathData$censor_state()}}
\item \href{#method-PathData-null_state}{\code{PathData$null_state()}}
\item \href{#method-PathData-state_names}{\code{PathData$state_names()}}
\item \href{#method-PathData-terminal_states}{\code{PathData$terminal_states()}}
\item \href{#method-PathData-get_event_states}{\code{PathData$get_event_states()}}
\item \href{#method-PathData-as_msdata}{\code{PathData$as_msdata()}}
\item \href{#method-PathData-as_time_to_first_event}{\code{PathData$as_time_to_first_event()}}
\item \href{#method-PathData-get_event_state_names}{\code{PathData$get_event_state_names()}}
\item \href{#method-PathData-print}{\code{PathData$print()}}
\item \href{#method-PathData-as_data_frame}{\code{PathData$as_data_frame()}}
\item \href{#method-PathData-as_transitions}{\code{PathData$as_transitions()}}
\item \href{#method-PathData-plot_paths}{\code{PathData$plot_paths()}}
\item \href{#method-PathData-trans_matrix}{\code{PathData$trans_matrix()}}
\item \href{#method-PathData-plot_graph}{\code{PathData$plot_graph()}}
\item \href{#method-PathData-coxph}{\code{PathData$coxph()}}
\item \href{#method-PathData-filter}{\code{PathData$filter()}}
\item \href{#method-PathData-clone}{\code{PathData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-new"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$new(
  subject_df,
  path_df,
  link_df,
  transmat,
  covs = NULL,
  check_order = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject_df}}{Data frame with one row per subject. Must have
\code{subject_id} and all covariates as columns.}

\item{\code{path_df}}{Data frame of actual paths. Must have \code{path_id},
\code{state}, \code{time}, \code{is_event}, \code{is_censor}
and \code{trans_idx} as columns.}

\item{\code{link_df}}{Links the path and subject data frames. Must have
\code{path_id}, \code{draw_idx}, \code{rep_idx}, and \code{subject_id} as
columns.}

\item{\code{transmat}}{A \code{\link{TransitionMatrix}} describing the system to
which the paths belong.}

\item{\code{covs}}{Covariate column names.}

\item{\code{check_order}}{Check order of paths?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-covariate_names"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-covariate_names}{}}}
\subsection{Method \code{covariate_names()}}{
Get names of covariates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$covariate_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-lengths"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-lengths}{}}}
\subsection{Method \code{lengths()}}{
Get path lenghts (among paths that include events)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$lengths()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a data frame of path ids and counts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-n_paths"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-n_paths}{}}}
\subsection{Method \code{n_paths()}}{
Get number of paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$n_paths()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-longest_path"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-longest_path}{}}}
\subsection{Method \code{longest_path()}}{
Get longest path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$longest_path()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{\link{PathData}} object with just one path
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-censor_state"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-censor_state}{}}}
\subsection{Method \code{censor_state()}}{
Get name of censoring state
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$censor_state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-null_state"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-null_state}{}}}
\subsection{Method \code{null_state()}}{
Get name of null state
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$null_state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-state_names"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-state_names}{}}}
\subsection{Method \code{state_names()}}{
Get names of all states, including censor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$state_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-terminal_states"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-terminal_states}{}}}
\subsection{Method \code{terminal_states()}}{
Get names of terminal states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$terminal_states()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-get_event_states"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-get_event_states}{}}}
\subsection{Method \code{get_event_states()}}{
Get indices of event states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$get_event_states()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_msdata"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_msdata}{}}}
\subsection{Method \code{as_msdata()}}{
Convert to format used by the 'mstate' package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_msdata(covariates = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Include covariates?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with elements \code{msdata} and
\code{legend}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_time_to_first_event"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_time_to_first_event}{}}}
\subsection{Method \code{as_time_to_first_event()}}{
Format as time to first event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_time_to_first_event(covs = c("subject_id"), truncate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covs}}{covariates to include in output data frame}

\item{\code{truncate}}{truncate after terminal events?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-get_event_state_names"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-get_event_state_names}{}}}
\subsection{Method \code{get_event_state_names()}}{
Get names of event states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$get_event_state_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-print"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-print}{}}}
\subsection{Method \code{print()}}{
Print info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_data_frame}{}}}
\subsection{Method \code{as_data_frame()}}{
Convert to one long data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_data_frame(covariates = NULL, truncate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{Which covariates to include?}

\item{\code{truncate}}{Truncate after terminal events?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-as_transitions"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-as_transitions}{}}}
\subsection{Method \code{as_transitions()}}{
Data frame in transitions format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$as_transitions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-plot_paths"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-plot_paths}{}}}
\subsection{Method \code{plot_paths()}}{
Step plot of the paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$plot_paths(n_paths = NULL, alpha = 0.5, truncate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_paths}}{Number of paths to subsample for plotting.}

\item{\code{alpha}}{opacity}

\item{\code{truncate}}{truncate after terminal events?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-trans_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-trans_matrix}{}}}
\subsection{Method \code{trans_matrix()}}{
Transition proportion matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$trans_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-plot_graph"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-plot_graph}{}}}
\subsection{Method \code{plot_graph()}}{
Visualize the transition proportion matrix as a graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$plot_graph(digits = 3, include_censor = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{Max number of digits to show in numbers}

\item{\code{include_censor}}{Include censoring state in the graph?}

\item{\code{...}}{Arguments passed to \code{qgraph}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{qgraph} plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-coxph"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-coxph}{}}}
\subsection{Method \code{coxph()}}{
Fit Cox proportional hazards model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$coxph(covs = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covs}}{Covariate names.}

\item{\code{...}}{Arguments passed to \code{survival::coxph}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-filter"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-filter}{}}}
\subsection{Method \code{filter()}}{
Filter based on path id, creates new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$filter(
  path_ids_keep = NULL,
  subject_ids_keep = NULL,
  rep_ids_keep = NULL,
  draw_ids_keep = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_ids_keep}}{Path ids to keep}

\item{\code{subject_ids_keep}}{Subject ids to keep}

\item{\code{rep_ids_keep}}{Repetition ids to keep}

\item{\code{draw_ids_keep}}{Draw ids to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PathData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PathData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
