% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKModel.R
\name{PKModel}
\alias{PKModel}
\title{One-compartment PK Model}
\description{
One-compartment PK Model

One-compartment PK Model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PKModel-new}{\code{PKModel$new()}}
\item \href{#method-PKModel-ka_covs}{\code{PKModel$ka_covs()}}
\item \href{#method-PKModel-CL_covs}{\code{PKModel$CL_covs()}}
\item \href{#method-PKModel-V2_covs}{\code{PKModel$V2_covs()}}
\item \href{#method-PKModel-covs}{\code{PKModel$covs()}}
\item \href{#method-PKModel-print}{\code{PKModel$print()}}
\item \href{#method-PKModel-simulate_ss}{\code{PKModel$simulate_ss()}}
\item \href{#method-PKModel-compute_ss_auc}{\code{PKModel$compute_ss_auc()}}
\item \href{#method-PKModel-simulate_data}{\code{PKModel$simulate_data()}}
\item \href{#method-PKModel-format_params}{\code{PKModel$format_params()}}
\item \href{#method-PKModel-clone}{\code{PKModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$new(covariates)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates}}{A list with elements \code{ka},
\code{CL}, and \code{V2}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-ka_covs"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-ka_covs}{}}}
\subsection{Method \code{ka_covs()}}{
Get the covariates affecting the k_a parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$ka_covs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-CL_covs"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-CL_covs}{}}}
\subsection{Method \code{CL_covs()}}{
Get the covariates affecting the CL parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$CL_covs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-V2_covs"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-V2_covs}{}}}
\subsection{Method \code{V2_covs()}}{
Get the covariates affecting the V2 parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$V2_covs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-covs"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-covs}{}}}
\subsection{Method \code{covs()}}{
Get names of all unique covariates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$covs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-print}{}}}
\subsection{Method \code{print()}}{
Print the object info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-simulate_ss"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-simulate_ss}{}}}
\subsection{Method \code{simulate_ss()}}{
Simulate system in steady state
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$simulate_ss(t, theta, dose, tau)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{output time points}

\item{\code{theta}}{parameter values}

\item{\code{dose}}{dose}

\item{\code{tau}}{dosing interval}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector with the same length as \code{t}, representing the
concentration in the central compartment.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-compute_ss_auc"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-compute_ss_auc}{}}}
\subsection{Method \code{compute_ss_auc()}}{
Compute steady-state area under curve over one dosing interval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$compute_ss_auc(theta, dose)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{parameter values}

\item{\code{dose}}{dose}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-simulate_data"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-simulate_data}{}}}
\subsection{Method \code{simulate_data()}}{
Simulate data with many subjects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$simulate_data(df_subjects, beta_pk = NULL, tau = 24, sigma = 0.2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_subjects}}{Data frame with one row for each subject}

\item{\code{beta_pk}}{Covariate effects}

\item{\code{tau}}{Dosing interval (PK)}

\item{\code{sigma}}{Noise magnitude of concentration measurements (PK)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with one row for each subject, and a
\code{\link{DosingData}} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-format_params"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-format_params}{}}}
\subsection{Method \code{format_params()}}{
Format list of input PK parameters to standardized format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$format_params(beta_pk = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{beta_pk}}{A list of max three elements}

\item{\code{return}}{A list with three elements}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PKModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PKModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PKModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
