% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathData.R
\name{potential_covariates}
\alias{potential_covariates}
\title{Look for potential covariates}
\usage{
potential_covariates(pd, possible = NULL, ...)
}
\arguments{
\item{pd}{A \code{\link{PathData}} object}

\item{possible}{Possible covariates to look for (character vector)}
}
\value{
A \code{data.frame}
}
\description{
Look for potential covariates
}
