% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DosingData.R
\name{simulate_dosing}
\alias{simulate_dosing}
\title{Simulate dosing data}
\usage{
simulate_dosing(df_subjects, tau = 24)
}
\arguments{
\item{df_subjects}{Data frame with one row for each subject}

\item{tau}{Dosing interval.}
}
\value{
A \code{\link{DosingData}} object
}
\description{
Simulate dosing data
}
