% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{simulate_example_data}
\alias{simulate_example_data}
\title{Simulate example data}
\usage{
simulate_example_data(N = 10, beta_haz = NULL, beta_pk = NULL, log_w0 = -4)
}
\arguments{
\item{N}{number of subjects}

\item{beta_haz}{see the documentation of \code{\link{MultistateModel}}}

\item{beta_pk}{see the documentation of \code{\link{MultistateModel}}}

\item{log_w0}{see the documentation of \code{\link{MultistateModel}}}
}
\description{
Simulate example data
}
