% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_time_evolution.R
\name{solve_time_evolution}
\alias{solve_time_evolution}
\title{Solve the Kolmogorov forward equation of a Markovian multistate system}
\usage{
solve_time_evolution(system, t, w, log_w0, log_m)
}
\arguments{
\item{system}{A \code{\link{MultistateSystem}}}

\item{t}{A time grid}

\item{w}{An array of shape \code{n_trans} x \code{n_weights}}

\item{log_w0}{A vector of length \code{n_trans}}

\item{log_m}{A vector of length \code{n_trans}}
}
\description{
Solve the Kolmogorov forward equation of a Markovian multistate system
}
