% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_time_evolution.R
\name{solve_trans_prob_matrix}
\alias{solve_trans_prob_matrix}
\title{Solve the transition probability matrix}
\usage{
solve_trans_prob_matrix(
  system,
  log_w0,
  w = NULL,
  log_m = NULL,
  t_init = 0,
  t_end = NULL
)
}
\arguments{
\item{system}{A \code{\link{MultistateSystem}}}

\item{log_w0}{A vector of length \code{n_trans}}

\item{w}{An array of shape \code{n_trans} x \code{n_weights}}

\item{log_m}{A vector of length \code{n_trans}}

\item{t_init}{Initial time}

\item{t_end}{End time}
}
\value{
A matrix \code{P} where \code{P[i,j]} is the probability that
the system will be in state \code{j} at time \code{t_end} given that it
is in state \code{i} at time \code{t_init}
}
\description{
Solve the transition probability matrix
}
