% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_time_evolution.R
\name{solve_trans_prob_matrix}
\alias{solve_trans_prob_matrix}
\title{Solve the transition probability matrix for each given output time}
\usage{
solve_trans_prob_matrix(
  system,
  t_out,
  log_w0,
  w = NULL,
  log_m = NULL,
  t_start = 0,
  ...
)
}
\arguments{
\item{system}{A \code{\link{MultistateSystem}}}

\item{t_out}{End times (vector)}

\item{log_w0}{A vector of length \code{n_trans}}

\item{w}{An array of shape \code{n_trans} x \code{n_weights}}

\item{log_m}{A vector of length \code{n_trans}}

\item{t_start}{Initial time}

\item{...}{Arguments passed to \code{deSolve::ode()}.}
}
\value{
An array \code{P} where \code{P[k,i,j]} is the probability that
the system will be in state \code{j} at the \code{k}th time point of
\code{t_out} given that it is in state \code{i} at time \code{t_start}.
}
\description{
Solve the transition probability matrix for each given output time
}
