% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_time_evolution.R
\name{solve_trans_prob_matrix_each_subject}
\alias{solve_trans_prob_matrix_each_subject}
\title{Solve transition probability matrices for each subject and draw in
'MultistateModelFit'}
\usage{
solve_trans_prob_matrix_each_subject(
  fit,
  t_start = 0,
  t_out = NULL,
  data = NULL,
  ...
)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{t_start}{Initial time}

\item{t_out}{End times (vector)}

\item{...}{Arguments passed to \code{deSolve::ode()}.}
}
\value{
A list with
\itemize{
  \item A 4-dimensional array \code{P} where \code{P[n,,,]} is the
    \code{P} matrix returned by \code{\link{solve_trans_prob_matrix}} for
    subject-draw combination \code{n}
   \item Index data frame
   \item The numeric vector \code{t_out}
 }
}
\description{
Solve transition probability matrices for each subject and draw in
'MultistateModelFit'
}
