% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_time_evolution.R
\name{solve_trans_prob_matrix_each_subject}
\alias{solve_trans_prob_matrix_each_subject}
\title{Solve transition probabilities for each subject in 'MultistateModelFit'}
\usage{
solve_trans_prob_matrix_each_subject(
  fit,
  t_start = 0,
  t_end = NULL,
  data = NULL
)
}
\arguments{
\item{fit}{A \code{\link{MultistateModelFit}} object}

\item{t_start}{Initial time.}

\item{t_end}{End time.}

\item{data}{A \code{\link{JointData}} object. If \code{NULL}, the
data used to fit the model is used. If not \code{NULL}, out-of-sample
mode is assumed (new subjects).}
}
\value{
For each subject and each draw, a matrix \code{P} where
\code{P[i,j]} is the probability that the system will be in state
\code{j} at time \code{t_end}
given that it is in state \code{i} at time \code{t_start}
}
\description{
Solve transition probabilities for each subject in 'MultistateModelFit'
}
