% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateGraph.R
\name{StateGraph}
\alias{StateGraph}
\title{Defines states and possible transitions}
\description{
Defines states and possible transitions

Defines states and possible transitions
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{matrix}}{a binary matrix}

\item{\code{states}}{a character vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StateGraph-new}{\code{StateGraph$new()}}
\item \href{#method-StateGraph-clone}{\code{StateGraph$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateGraph-new"></a>}}
\if{latex}{\out{\hypertarget{method-StateGraph-new}{}}}
\subsection{Method \code{new()}}{
Create graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateGraph$new(matrix, states)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{Binary square matrix that defines the states and possible}

\item{\code{states}}{Names of the states
transitions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateGraph-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StateGraph-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateGraph$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
