% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{TransitionMatrix}
\alias{TransitionMatrix}
\title{Defines states and possible transitions}
\description{
Defines states and possible transitions

Defines states and possible transitions
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{matrix}}{a binary \code{N} x \code{N} matrix where
\code{matrix[i,j]} is 1 if transition
from state \code{i} to \code{j} is possible}

\item{\code{states}}{a character vector of state names (length \code{N})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransitionMatrix-new}{\code{TransitionMatrix$new()}}
\item \href{#method-TransitionMatrix-num_states}{\code{TransitionMatrix$num_states()}}
\item \href{#method-TransitionMatrix-num_trans}{\code{TransitionMatrix$num_trans()}}
\item \href{#method-TransitionMatrix-num_trans_types}{\code{TransitionMatrix$num_trans_types()}}
\item \href{#method-TransitionMatrix-as_matrix}{\code{TransitionMatrix$as_matrix()}}
\item \href{#method-TransitionMatrix-as_transition_index_matrix}{\code{TransitionMatrix$as_transition_index_matrix()}}
\item \href{#method-TransitionMatrix-as_mstate_transmat}{\code{TransitionMatrix$as_mstate_transmat()}}
\item \href{#method-TransitionMatrix-possible_transitions_from}{\code{TransitionMatrix$possible_transitions_from()}}
\item \href{#method-TransitionMatrix-target_state}{\code{TransitionMatrix$target_state()}}
\item \href{#method-TransitionMatrix-at_risk}{\code{TransitionMatrix$at_risk()}}
\item \href{#method-TransitionMatrix-states_df}{\code{TransitionMatrix$states_df()}}
\item \href{#method-TransitionMatrix-trans_df}{\code{TransitionMatrix$trans_df()}}
\item \href{#method-TransitionMatrix-plot}{\code{TransitionMatrix$plot()}}
\item \href{#method-TransitionMatrix-print}{\code{TransitionMatrix$print()}}
\item \href{#method-TransitionMatrix-absorbing_states}{\code{TransitionMatrix$absorbing_states()}}
\item \href{#method-TransitionMatrix-source_states}{\code{TransitionMatrix$source_states()}}
\item \href{#method-TransitionMatrix-clone}{\code{TransitionMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-new"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-new}{}}}
\subsection{Method \code{new()}}{
Create graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$new(matrix, states)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{a binary \code{N} x \code{N} matrix where
\code{matrix[i,j]} is 1 if transition
from state \code{i} to \code{j} is possible}

\item{\code{states}}{a character vector of state names (length \code{N})
Number of states}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-num_states"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-num_states}{}}}
\subsection{Method \code{num_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$num_states()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer \code{N}
Number of transitions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-num_trans"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-num_trans}{}}}
\subsection{Method \code{num_trans()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$num_trans()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-num_trans_types"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-num_trans_types}{}}}
\subsection{Method \code{num_trans_types()}}{
Get number of different transition types.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$num_trans_types()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-as_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-as_matrix}{}}}
\subsection{Method \code{as_matrix()}}{
As a matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$as_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An \code{N} x \code{N} matrix with named columns and rows
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-as_transition_index_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-as_transition_index_matrix}{}}}
\subsection{Method \code{as_transition_index_matrix()}}{
As a matrix indexing the transitions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$as_transition_index_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An \code{N} x \code{N} matrix with named columns and rows
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-as_mstate_transmat"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-as_mstate_transmat}{}}}
\subsection{Method \code{as_mstate_transmat()}}{
Transition index matrix to format used by 'mstate'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$as_mstate_transmat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-possible_transitions_from"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-possible_transitions_from}{}}}
\subsection{Method \code{possible_transitions_from()}}{
Get indices of possible transitions from given state
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$possible_transitions_from(state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{Index of state
Get target state of a transition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-target_state"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-target_state}{}}}
\subsection{Method \code{target_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$target_state(trans_idx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trans_idx}}{Index of the transition
Get states that are at risk when at given state}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-at_risk"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-at_risk}{}}}
\subsection{Method \code{at_risk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$at_risk(state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{index of current state}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
indices of states at risk
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-states_df"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-states_df}{}}}
\subsection{Method \code{states_df()}}{
A data frame of states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$states_df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-trans_df"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-trans_df}{}}}
\subsection{Method \code{trans_df()}}{
A data frame of transitions ("legend")
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$trans_df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-plot}{}}}
\subsection{Method \code{plot()}}{
Visualize the matrix as a graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$plot(edge_labs = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{edge_labs}}{Edge labels}

\item{\code{...}}{Arguments passed to \code{qgraph::qgraph}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{qgraph} plot
Print output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-print"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
Get states that cannot be transitioned from
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-absorbing_states"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-absorbing_states}{}}}
\subsection{Method \code{absorbing_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$absorbing_states(names = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Return names of the states? Otherwise returns indices.
Get states that cannot be transitioned to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-source_states"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-source_states}{}}}
\subsection{Method \code{source_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$source_states(names = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Return names of the states? Otherwise returns indices.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
