% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\name{TransitionMatrix}
\alias{TransitionMatrix}
\title{Defines states and possible transitions}
\description{
Defines states and possible transitions

Defines states and possible transitions
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{matrix}}{a binary \code{N} x \code{N} matrix where
\code{matrix[i,j]} is 1 if transition
from state \code{i} to \code{j} is possible}

\item{\code{states}}{a character vector of state names (length \code{N})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransitionMatrix-new}{\code{TransitionMatrix$new()}}
\item \href{#method-TransitionMatrix-num_states}{\code{TransitionMatrix$num_states()}}
\item \href{#method-TransitionMatrix-as_matrix}{\code{TransitionMatrix$as_matrix()}}
\item \href{#method-TransitionMatrix-as_transition_index_matrix}{\code{TransitionMatrix$as_transition_index_matrix()}}
\item \href{#method-TransitionMatrix-states_df}{\code{TransitionMatrix$states_df()}}
\item \href{#method-TransitionMatrix-plot}{\code{TransitionMatrix$plot()}}
\item \href{#method-TransitionMatrix-print}{\code{TransitionMatrix$print()}}
\item \href{#method-TransitionMatrix-absorbing_states}{\code{TransitionMatrix$absorbing_states()}}
\item \href{#method-TransitionMatrix-source_states}{\code{TransitionMatrix$source_states()}}
\item \href{#method-TransitionMatrix-clone}{\code{TransitionMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-new"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-new}{}}}
\subsection{Method \code{new()}}{
Create graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$new(matrix, states)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{a binary \code{N} x \code{N} matrix where
\code{matrix[i,j]} is 1 if transition
from state \code{i} to \code{j} is possible}

\item{\code{states}}{a character vector of state names (length \code{N})
Number of states}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-num_states"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-num_states}{}}}
\subsection{Method \code{num_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$num_states()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer \code{N}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-as_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-as_matrix}{}}}
\subsection{Method \code{as_matrix()}}{
As a matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$as_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An \code{N} x \code{N} matrix with named columns and rows
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-as_transition_index_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-as_transition_index_matrix}{}}}
\subsection{Method \code{as_transition_index_matrix()}}{
As a matrix indexing the transitions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$as_transition_index_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An \code{N} x \code{N} matrix with named columns and rows
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-states_df"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-states_df}{}}}
\subsection{Method \code{states_df()}}{
A data frame of states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$states_df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{data.frame} with columns \code{state_idx} and \code{state}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-plot}{}}}
\subsection{Method \code{plot()}}{
Visualize the matrix as a graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \code{qgraph}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{qgraph} plot
Print output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-print"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
Get states that cannot be transitioned from
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-absorbing_states"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-absorbing_states}{}}}
\subsection{Method \code{absorbing_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$absorbing_states(names = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Return names of the states? Otherwise returns indices.
Get states that cannot be transitioned to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-source_states"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-source_states}{}}}
\subsection{Method \code{source_states()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$source_states(names = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Return names of the states? Otherwise returns indices.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
